### 2022-06-14 #4473 @arfon
Note for other editors, this is a re-submission of a previously-rejecting submission (https://github.com/openjournals/joss-reviews/issues/4287)
