+++ README.md
@@ -143,36 +143,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-06-13 13:32:47] L2 sensitivity is: 0.1281
+#> [2022-06-20 13:34:42] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-06-13 13:32:48] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-06-20 13:34:43] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-06-13 13:32:48] Initializing ROC-GLM
+#> [2022-06-20 13:34:43] Initializing ROC-GLM
 #> 
-#> [2022-06-13 13:32:48] Host: Received scores of negative response
-#> [2022-06-13 13:32:48] Receiving negative scores
-#> [2022-06-13 13:32:49] Host: Pushing pooled scores
-#> [2022-06-13 13:32:49] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-06-13 13:32:50] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-06-13 13:32:51] Deviance of iter1=137.2431
-#> [2022-06-13 13:32:52] Deviance of iter2=121.5994
-#> [2022-06-13 13:32:53] Deviance of iter3=147.7237
-#> [2022-06-13 13:32:54] Deviance of iter4=140.4008
-#> [2022-06-13 13:32:55] Deviance of iter5=129.2244
-#> [2022-06-13 13:32:56] Deviance of iter6=123.9979
-#> [2022-06-13 13:32:57] Deviance of iter7=123.1971
-#> [2022-06-13 13:32:58] Deviance of iter8=124.1615
-#> [2022-06-13 13:32:58] Deviance of iter9=124.5356
-#> [2022-06-13 13:32:59] Deviance of iter10=124.5503
-#> [2022-06-13 13:33:00] Deviance of iter11=124.5504
-#> [2022-06-13 13:33:01] Deviance of iter12=124.5504
-#> [2022-06-13 13:33:01] Host: Finished calculating ROC-GLM
-#> [2022-06-13 13:33:01] Host: Cleaning data on server
-#> [2022-06-13 13:33:02] Host: Calculating AUC and CI
-#> [2022-06-13 13:33:10] Finished!
+#> [2022-06-20 13:34:43] Host: Received scores of negative response
+#> [2022-06-20 13:34:43] Receiving negative scores
+#> [2022-06-20 13:34:44] Host: Pushing pooled scores
+#> [2022-06-20 13:34:45] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-06-20 13:34:45] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-06-20 13:34:46] Deviance of iter1=137.2431
+#> [2022-06-20 13:34:47] Deviance of iter2=121.5994
+#> [2022-06-20 13:34:48] Deviance of iter3=147.7237
+#> [2022-06-20 13:34:49] Deviance of iter4=140.4008
+#> [2022-06-20 13:34:50] Deviance of iter5=129.2244
+#> [2022-06-20 13:34:51] Deviance of iter6=123.9979
+#> [2022-06-20 13:34:52] Deviance of iter7=123.1971
+#> [2022-06-20 13:34:53] Deviance of iter8=124.1615
[…]
