+++ README.md
@@ -87,14 +87,17 @@ builder$append(
 connections = datashield.login(logins = builder$build(), assign = TRUE)
 #> 
 #> Logging into the collaborating servers
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Login ds1 [========================>---------------------------------------------------]  33% / 0s  Login ds2 [==================================================>-------------------------]  67% / 0s  Logged in all servers [================================================================] 100% / 1s
 ```
 
 #### Assign iris and validation vector at DataSHIELD (just for testing)
 
 ``` r
 datashield.assign(connections, "iris", quote(iris))
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (iris <- iris) [----------------------------------------------------------]   0% / 1s  Finalizing assignment ds1 (iris <- iris) [==============>------------------------------]  33% / 1s  Checking ds2 (iris <- iris) [==================>---------------------------------------]  33% / 1s  Finalizing assignment ds2 (iris <- iris) [=============================>---------------]  67% / 1s  Assigned expr. (iris <- iris) [========================================================] 100% / 1s
 vcall = paste0("quote(c(", paste(rep(c(1, 0), times = c(50, 100)), collapse = ", "), "))")
 datashield.assign(connections, "y", eval(parse(text = vcall)))
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (y <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ) [---]   0% / 0s  Finalizing assignment ds1 (y <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, )...  Checking ds2 (y <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ) [>--]  33% / 0s  Finalizing assignment ds2 (y <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, )...  Assigned expr. (y <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ) [=] 100% / 0s
 ```
 
 #### Load test model, push to DataSHIELD, and calculate predictions
@@ -106,10 +109,12 @@ mod = glm(y ~ Sepal.Length, data = iris, family = binomial())
 
 # Push the model to the DataSHIELD servers using `dsPredictBase`:
 pushObject(connections, mod)
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (mod <- decodeBinary("580a000000030004020000030500000000055554462d3800000313000000...  Finalizing assignment ds1 (mod <- decodeBinary("580a000000030004020000030500000000055554462d380...  Checking ds2 (mod <- decodeBinary("580a000000030004020000030500000000055554462d3800000313000000...  Finalizing assignment ds2 (mod <- decodeBinary("580a000000030004020000030500000000055554462d380...  Assigned expr. (mod <- decodeBinary("580a000000030004020000030500000000055554462d38000003130000...
 
 # Calculate scores and save at the servers using `dsPredictBase`:
 pfun =  "predict(mod, newdata = D, type = 'response')"
[…]
