### 2022-06-24 #4473 @schalkdaniel
@csoneson - Sure, the first idea we had was to have multiple `R` packages to evaluate a binary classification model in a distributed setting. Therefore, we had a package for ROC curve and AUC calculation, calibration curves and other metrics, and a package for basic stuff like distributing an `R` model to multiple [DataSHIELD](https://www.datashield.org/) servers. The previous submission just covers the ROC curve and AUC calculation while this submission is a merge and update of the three packages. So it contains the functionality to evaluate statistical or predictive models in multiple ways instead of just calculating the ROC curve and the AUC.

I don't really have preferences for the reviewer. But, I picked three candidates from the linked list who seem active on GitHub and do have a machine learning / biostats / bioinformatics background:
- zhiiiyang
- jaybee84
- brunomontezano
