### 2022-06-25 #4473 @csoneson
@schalkdaniel - great, thanks for the summary!

👋🏻 @zhiiiyang, @brunomontezano - would you be interested in reviewing this submission for JOSS?

> dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD
