+++ README.md
@@ -143,36 +143,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-06-24 07:09:55] L2 sensitivity is: 0.1281
+#> [2022-06-27 13:38:29] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-06-24 07:09:56] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-06-27 13:38:30] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-06-24 07:09:56] Initializing ROC-GLM
+#> [2022-06-27 13:38:30] Initializing ROC-GLM
 #> 
-#> [2022-06-24 07:09:56] Host: Received scores of negative response
-#> [2022-06-24 07:09:56] Receiving negative scores
-#> [2022-06-24 07:09:57] Host: Pushing pooled scores
-#> [2022-06-24 07:09:59] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-06-24 07:10:00] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-06-24 07:10:02] Deviance of iter1=137.2431
-#> [2022-06-24 07:10:03] Deviance of iter2=121.5994
-#> [2022-06-24 07:10:04] Deviance of iter3=147.7237
-#> [2022-06-24 07:10:05] Deviance of iter4=140.4008
-#> [2022-06-24 07:10:07] Deviance of iter5=129.2244
-#> [2022-06-24 07:10:08] Deviance of iter6=123.9979
-#> [2022-06-24 07:10:09] Deviance of iter7=123.1971
-#> [2022-06-24 07:10:11] Deviance of iter8=124.1615
-#> [2022-06-24 07:10:12] Deviance of iter9=124.5356
-#> [2022-06-24 07:10:13] Deviance of iter10=124.5503
-#> [2022-06-24 07:10:15] Deviance of iter11=124.5504
-#> [2022-06-24 07:10:16] Deviance of iter12=124.5504
-#> [2022-06-24 07:10:16] Host: Finished calculating ROC-GLM
-#> [2022-06-24 07:10:16] Host: Cleaning data on server
-#> [2022-06-24 07:10:17] Host: Calculating AUC and CI
-#> [2022-06-24 07:10:29] Finished!
+#> [2022-06-27 13:38:30] Host: Received scores of negative response
+#> [2022-06-27 13:38:30] Receiving negative scores
+#> [2022-06-27 13:38:31] Host: Pushing pooled scores
+#> [2022-06-27 13:38:32] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-06-27 13:38:33] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-06-27 13:38:34] Deviance of iter1=137.2431
+#> [2022-06-27 13:38:35] Deviance of iter2=121.5994
+#> [2022-06-27 13:38:36] Deviance of iter3=147.7237
+#> [2022-06-27 13:38:37] Deviance of iter4=140.4008
+#> [2022-06-27 13:38:37] Deviance of iter5=129.2244
+#> [2022-06-27 13:38:38] Deviance of iter6=123.9979
+#> [2022-06-27 13:38:39] Deviance of iter7=123.1971
+#> [2022-06-27 13:38:40] Deviance of iter8=124.1615
+#> [2022-06-27 13:38:41] Deviance of iter9=124.5356
[…]
