+++ README.md
@@ -45,6 +45,31 @@ publishing the package on OPAL (see
 Note that the package needs to be installed at both locations, the
 server and the analysts machine.
 
+## Installation on DataSHIELD
+
+The two options are to use the Opal API: - Log into Opal ans switch to
+the `Administration/DataSHIELD/` tab - Click the `Add DataSHIELD
+package` button - Select `GitHub` as source, and use `difuture-lmu` as
+user, `dsBinVal` as name, and `main` as Git reference.
+
+The second option is to use the `opalr` package to install `dsBinVal`
+directly from `R`:
+
+``` r
+### User credentials (here from the opal test server):
+surl     = "https://opal-demo.obiba.org/"
+username = "administrator"
+password = "password"
+
+### Install package and publish methods:
+opal = opalr::opal.login(username = username, password = password, url = surl)
+
+opalr::dsadmin.install_github_package(opal = opal, pkg = "dsBinVal", username = "difuture-lmu", ref = "main")
+opalr::dsadmin.publish_package(opal = opal, pkg = "dsBinVal")
+
+opalr::opal.logout(opal)
+```
+
 ## Usage
 
 A more sophisticated example is available
@@ -143,36 +168,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-06-27 13:38:29] L2 sensitivity is: 0.1281
+#> [2022-06-28 07:32:32] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-06-27 13:38:30] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-06-28 07:32:33] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-06-27 13:38:30] Initializing ROC-GLM
+#> [2022-06-28 07:32:33] Initializing ROC-GLM
 #> 
-#> [2022-06-27 13:38:30] Host: Received scores of negative response
-#> [2022-06-27 13:38:30] Receiving negative scores
-#> [2022-06-27 13:38:31] Host: Pushing pooled scores
-#> [2022-06-27 13:38:32] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-06-27 13:38:33] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-06-27 13:38:34] Deviance of iter1=137.2431
-#> [2022-06-27 13:38:35] Deviance of iter2=121.5994
-#> [2022-06-27 13:38:36] Deviance of iter3=147.7237
-#> [2022-06-27 13:38:37] Deviance of iter4=140.4008
-#> [2022-06-27 13:38:37] Deviance of iter5=129.2244
-#> [2022-06-27 13:38:38] Deviance of iter6=123.9979
-#> [2022-06-27 13:38:39] Deviance of iter7=123.1971
-#> [2022-06-27 13:38:40] Deviance of iter8=124.1615
-#> [2022-06-27 13:38:41] Deviance of iter9=124.5356
-#> [2022-06-27 13:38:42] Deviance of iter10=124.5503
-#> [2022-06-27 13:38:43] Deviance of iter11=124.5504
-#> [2022-06-27 13:38:44] Deviance of iter12=124.5504
[…]
