+++ README.Rmd
@@ -60,6 +60,7 @@ Note that the package needs to be installed at both locations, the server and th
 ## Installation on DataSHIELD
 
 The two options are to use the Opal API:
+
 - Log into Opal ans switch to the `Administration/DataSHIELD/` tab
 - Click the `Add DataSHIELD package` button
 - Select `GitHub` as source, and use `difuture-lmu` as user, `dsBinVal` as name, and `main` as Git reference.
