### 2022-07-03 #4473 @csoneson
👋🏻 @gsfyrakis, @patRyserWelch8, @MaximLippeveld - would (two of) you be available and interested in reviewing this submission for JOSS?

> dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD
