### 2022-07-04 #4473 @patRyserWelch8
HI

Sure. Always happy to help.  For when do you need it?

Best wishes,

P.

Dr Pat Ryser-Welch

Staff tutor/Associate lecturer

Referred as: Her/She

Calendar: up to date, so feel free to send me a calendar invite, if you would benefit from my contributions

Let's remember:

Start by doing what's necessary; then do what's possible; and suddenly you are doing the impossible.

 Francis of Assisi

Mankind remains the most extraordinary computer of all....

John F. Kennedy

Read me : https://interestingcomputing.com
