### 2022-07-04 #4473 @csoneson
@patRyserWelch8 Thank you! We usually ask reviewers to provide a first round of reviews within ~2 weeks (once we have secured two reviewers, I will assign you and open the review issue, where you will be able to generate your review checklist and start the review). However, I should also note that in many cases  it is easier to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
