### 2022-07-05 #4473 @brunomontezano
> @schalkdaniel - great, thanks for the summary!
>
> 👋🏻 @zhiiiyang, @brunomontezano - would you be interested in reviewing this submission for JOSS?
>
> > dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD

Firstly, thanks for the invitation. And secondly, I accept the invite to review this paper.
