### 2022-07-05 #4545 @editorialbot
[REVIEW]: dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD
**Submitting author:** @schalkdaniel (<a href="http://orcid.org/0000-0003-0950-1947">Daniel Schalk</a>)
**Repository:** https://github.com/difuture-lmu/dsBinVal
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2
