+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-07-05 14:26:01] L2 sensitivity is: 0.1281
+#> [2022-07-11 13:48:04] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-07-05 14:26:02] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-07-11 13:48:05] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-07-05 14:26:02] Initializing ROC-GLM
+#> [2022-07-11 13:48:05] Initializing ROC-GLM
 #> 
-#> [2022-07-05 14:26:02] Host: Received scores of negative response
-#> [2022-07-05 14:26:02] Receiving negative scores
-#> [2022-07-05 14:26:03] Host: Pushing pooled scores
-#> [2022-07-05 14:26:05] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-07-05 14:26:06] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-07-05 14:26:07] Deviance of iter1=137.2431
-#> [2022-07-05 14:26:09] Deviance of iter2=121.5994
-#> [2022-07-05 14:26:10] Deviance of iter3=147.7237
-#> [2022-07-05 14:26:11] Deviance of iter4=140.4008
-#> [2022-07-05 14:26:12] Deviance of iter5=129.2244
-#> [2022-07-05 14:26:14] Deviance of iter6=123.9979
-#> [2022-07-05 14:26:15] Deviance of iter7=123.1971
-#> [2022-07-05 14:26:16] Deviance of iter8=124.1615
-#> [2022-07-05 14:26:17] Deviance of iter9=124.5356
-#> [2022-07-05 14:26:19] Deviance of iter10=124.5503
-#> [2022-07-05 14:26:20] Deviance of iter11=124.5504
-#> [2022-07-05 14:26:21] Deviance of iter12=124.5504
-#> [2022-07-05 14:26:21] Host: Finished calculating ROC-GLM
-#> [2022-07-05 14:26:21] Host: Cleaning data on server
-#> [2022-07-05 14:26:22] Host: Calculating AUC and CI
-#> [2022-07-05 14:26:34] Finished!
+#> [2022-07-11 13:48:05] Host: Received scores of negative response
+#> [2022-07-11 13:48:05] Receiving negative scores
+#> [2022-07-11 13:48:06] Host: Pushing pooled scores
+#> [2022-07-11 13:48:08] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-07-11 13:48:09] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-07-11 13:48:11] Deviance of iter1=137.2431
+#> [2022-07-11 13:48:12] Deviance of iter2=121.5994
+#> [2022-07-11 13:48:13] Deviance of iter3=147.7237
+#> [2022-07-11 13:48:15] Deviance of iter4=140.4008
+#> [2022-07-11 13:48:16] Deviance of iter5=129.2244
+#> [2022-07-11 13:48:18] Deviance of iter6=123.9979
+#> [2022-07-11 13:48:19] Deviance of iter7=123.1971
+#> [2022-07-11 13:48:20] Deviance of iter8=124.1615
[…]
