+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-07-18 13:48:52] L2 sensitivity is: 0.1281
+#> [2022-07-25 13:37:01] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-07-18 13:48:54] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-07-25 13:37:02] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-07-18 13:48:54] Initializing ROC-GLM
+#> [2022-07-25 13:37:02] Initializing ROC-GLM
 #> 
-#> [2022-07-18 13:48:54] Host: Received scores of negative response
-#> [2022-07-18 13:48:54] Receiving negative scores
-#> [2022-07-18 13:48:55] Host: Pushing pooled scores
-#> [2022-07-18 13:48:56] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-07-18 13:48:58] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-07-18 13:48:59] Deviance of iter1=137.2431
-#> [2022-07-18 13:49:00] Deviance of iter2=121.5994
-#> [2022-07-18 13:49:02] Deviance of iter3=147.7237
-#> [2022-07-18 13:49:03] Deviance of iter4=140.4008
-#> [2022-07-18 13:49:04] Deviance of iter5=129.2244
-#> [2022-07-18 13:49:06] Deviance of iter6=123.9979
-#> [2022-07-18 13:49:07] Deviance of iter7=123.1971
-#> [2022-07-18 13:49:09] Deviance of iter8=124.1615
-#> [2022-07-18 13:49:10] Deviance of iter9=124.5356
-#> [2022-07-18 13:49:11] Deviance of iter10=124.5503
-#> [2022-07-18 13:49:13] Deviance of iter11=124.5504
-#> [2022-07-18 13:49:14] Deviance of iter12=124.5504
-#> [2022-07-18 13:49:14] Host: Finished calculating ROC-GLM
-#> [2022-07-18 13:49:14] Host: Cleaning data on server
-#> [2022-07-18 13:49:15] Host: Calculating AUC and CI
-#> [2022-07-18 13:49:27] Finished!
+#> [2022-07-25 13:37:02] Host: Received scores of negative response
+#> [2022-07-25 13:37:02] Receiving negative scores
+#> [2022-07-25 13:37:03] Host: Pushing pooled scores
+#> [2022-07-25 13:37:04] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-07-25 13:37:05] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-07-25 13:37:06] Deviance of iter1=137.2431
+#> [2022-07-25 13:37:06] Deviance of iter2=121.5994
+#> [2022-07-25 13:37:07] Deviance of iter3=147.7237
+#> [2022-07-25 13:37:08] Deviance of iter4=140.4008
+#> [2022-07-25 13:37:09] Deviance of iter5=129.2244
+#> [2022-07-25 13:37:10] Deviance of iter6=123.9979
+#> [2022-07-25 13:37:11] Deviance of iter7=123.1971
+#> [2022-07-25 13:37:12] Deviance of iter8=124.1615
[…]
