+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-07-25 13:37:01] L2 sensitivity is: 0.1281
+#> [2022-08-01 13:39:00] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-07-25 13:37:02] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-08-01 13:39:01] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-07-25 13:37:02] Initializing ROC-GLM
+#> [2022-08-01 13:39:01] Initializing ROC-GLM
 #> 
-#> [2022-07-25 13:37:02] Host: Received scores of negative response
-#> [2022-07-25 13:37:02] Receiving negative scores
-#> [2022-07-25 13:37:03] Host: Pushing pooled scores
-#> [2022-07-25 13:37:04] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-07-25 13:37:05] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-07-25 13:37:06] Deviance of iter1=137.2431
-#> [2022-07-25 13:37:06] Deviance of iter2=121.5994
-#> [2022-07-25 13:37:07] Deviance of iter3=147.7237
-#> [2022-07-25 13:37:08] Deviance of iter4=140.4008
-#> [2022-07-25 13:37:09] Deviance of iter5=129.2244
-#> [2022-07-25 13:37:10] Deviance of iter6=123.9979
-#> [2022-07-25 13:37:11] Deviance of iter7=123.1971
-#> [2022-07-25 13:37:12] Deviance of iter8=124.1615
-#> [2022-07-25 13:37:12] Deviance of iter9=124.5356
-#> [2022-07-25 13:37:13] Deviance of iter10=124.5503
-#> [2022-07-25 13:37:14] Deviance of iter11=124.5504
-#> [2022-07-25 13:37:15] Deviance of iter12=124.5504
-#> [2022-07-25 13:37:15] Host: Finished calculating ROC-GLM
-#> [2022-07-25 13:37:15] Host: Cleaning data on server
-#> [2022-07-25 13:37:16] Host: Calculating AUC and CI
-#> [2022-07-25 13:37:24] Finished!
+#> [2022-08-01 13:39:01] Host: Received scores of negative response
+#> [2022-08-01 13:39:01] Receiving negative scores
+#> [2022-08-01 13:39:02] Host: Pushing pooled scores
+#> [2022-08-01 13:39:03] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-08-01 13:39:03] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-08-01 13:39:04] Deviance of iter1=137.2431
+#> [2022-08-01 13:39:05] Deviance of iter2=121.5994
+#> [2022-08-01 13:39:06] Deviance of iter3=147.7237
+#> [2022-08-01 13:39:07] Deviance of iter4=140.4008
+#> [2022-08-01 13:39:07] Deviance of iter5=129.2244
+#> [2022-08-01 13:39:08] Deviance of iter6=123.9979
+#> [2022-08-01 13:39:09] Deviance of iter7=123.1971
+#> [2022-08-01 13:39:10] Deviance of iter8=124.1615
[…]
