+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-08-01 13:39:00] L2 sensitivity is: 0.1281
+#> [2022-08-08 13:38:40] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-08-01 13:39:01] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-08-08 13:38:41] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-08-01 13:39:01] Initializing ROC-GLM
+#> [2022-08-08 13:38:41] Initializing ROC-GLM
 #> 
-#> [2022-08-01 13:39:01] Host: Received scores of negative response
-#> [2022-08-01 13:39:01] Receiving negative scores
-#> [2022-08-01 13:39:02] Host: Pushing pooled scores
-#> [2022-08-01 13:39:03] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-08-01 13:39:03] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-08-01 13:39:04] Deviance of iter1=137.2431
-#> [2022-08-01 13:39:05] Deviance of iter2=121.5994
-#> [2022-08-01 13:39:06] Deviance of iter3=147.7237
-#> [2022-08-01 13:39:07] Deviance of iter4=140.4008
-#> [2022-08-01 13:39:07] Deviance of iter5=129.2244
-#> [2022-08-01 13:39:08] Deviance of iter6=123.9979
-#> [2022-08-01 13:39:09] Deviance of iter7=123.1971
-#> [2022-08-01 13:39:10] Deviance of iter8=124.1615
-#> [2022-08-01 13:39:11] Deviance of iter9=124.5356
-#> [2022-08-01 13:39:11] Deviance of iter10=124.5503
-#> [2022-08-01 13:39:12] Deviance of iter11=124.5504
-#> [2022-08-01 13:39:13] Deviance of iter12=124.5504
-#> [2022-08-01 13:39:13] Host: Finished calculating ROC-GLM
-#> [2022-08-01 13:39:13] Host: Cleaning data on server
-#> [2022-08-01 13:39:14] Host: Calculating AUC and CI
-#> [2022-08-01 13:39:21] Finished!
+#> [2022-08-08 13:38:41] Host: Received scores of negative response
+#> [2022-08-08 13:38:41] Receiving negative scores
+#> [2022-08-08 13:38:42] Host: Pushing pooled scores
+#> [2022-08-08 13:38:43] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-08-08 13:38:44] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-08-08 13:38:45] Deviance of iter1=137.2431
+#> [2022-08-08 13:38:45] Deviance of iter2=121.5994
+#> [2022-08-08 13:38:46] Deviance of iter3=147.7237
+#> [2022-08-08 13:38:47] Deviance of iter4=140.4008
+#> [2022-08-08 13:38:48] Deviance of iter5=129.2244
+#> [2022-08-08 13:38:49] Deviance of iter6=123.9979
+#> [2022-08-08 13:38:50] Deviance of iter7=123.1971
+#> [2022-08-08 13:38:51] Deviance of iter8=124.1615
[…]
