### 2022-08-10 #4545 @brunomontezano
@schalkdaniel Some suggestions about the paper:

- Add _State of the field_ section. Maybe you could briefly describe some other technologies that aim to solve the same problem and how the interaction with DataSHIELD framework in R could be different in usage than these alternatives. If it would ease or improve the current approach, etc.

About the package repository itself:

- I'd recommend you to add some community guidelines, so, how other people could contribute to the software (code-related or not), how issues or bugs could be reported (email? issue tab on GitHub?), how people could seek help about the software usage (issue tab? email? self-hosted platform?). It could be included in the end of README if you like.

Feel free not to accept my suggestions as long as you have good arguments. Any questions, just ask me.
