+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-08-08 13:38:40] L2 sensitivity is: 0.1281
+#> [2022-08-15 13:37:11] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-08-08 13:38:41] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-08-15 13:37:12] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-08-08 13:38:41] Initializing ROC-GLM
+#> [2022-08-15 13:37:12] Initializing ROC-GLM
 #> 
-#> [2022-08-08 13:38:41] Host: Received scores of negative response
-#> [2022-08-08 13:38:41] Receiving negative scores
-#> [2022-08-08 13:38:42] Host: Pushing pooled scores
-#> [2022-08-08 13:38:43] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-08-08 13:38:44] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-08-08 13:38:45] Deviance of iter1=137.2431
-#> [2022-08-08 13:38:45] Deviance of iter2=121.5994
-#> [2022-08-08 13:38:46] Deviance of iter3=147.7237
-#> [2022-08-08 13:38:47] Deviance of iter4=140.4008
-#> [2022-08-08 13:38:48] Deviance of iter5=129.2244
-#> [2022-08-08 13:38:49] Deviance of iter6=123.9979
-#> [2022-08-08 13:38:50] Deviance of iter7=123.1971
-#> [2022-08-08 13:38:51] Deviance of iter8=124.1615
-#> [2022-08-08 13:38:52] Deviance of iter9=124.5356
-#> [2022-08-08 13:38:52] Deviance of iter10=124.5503
-#> [2022-08-08 13:38:53] Deviance of iter11=124.5504
-#> [2022-08-08 13:38:54] Deviance of iter12=124.5504
-#> [2022-08-08 13:38:54] Host: Finished calculating ROC-GLM
-#> [2022-08-08 13:38:54] Host: Cleaning data on server
-#> [2022-08-08 13:38:55] Host: Calculating AUC and CI
-#> [2022-08-08 13:39:03] Finished!
+#> [2022-08-15 13:37:12] Host: Received scores of negative response
+#> [2022-08-15 13:37:12] Receiving negative scores
+#> [2022-08-15 13:37:13] Host: Pushing pooled scores
+#> [2022-08-15 13:37:14] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-08-15 13:37:14] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-08-15 13:37:15] Deviance of iter1=137.2431
+#> [2022-08-15 13:37:16] Deviance of iter2=121.5994
+#> [2022-08-15 13:37:17] Deviance of iter3=147.7237
+#> [2022-08-15 13:37:18] Deviance of iter4=140.4008
+#> [2022-08-15 13:37:19] Deviance of iter5=129.2244
+#> [2022-08-15 13:37:20] Deviance of iter6=123.9979
+#> [2022-08-15 13:37:20] Deviance of iter7=123.1971
+#> [2022-08-15 13:37:21] Deviance of iter8=124.1615
[…]
