### 2022-08-27 0c056c5
Update R-CMD-check.yaml
Remove Ubuntu 18-04
