+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-08-22 13:39:18] L2 sensitivity is: 0.1281
+#> [2022-08-27 13:20:26] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-08-22 13:39:19] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-08-27 13:20:27] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-08-22 13:39:19] Initializing ROC-GLM
+#> [2022-08-27 13:20:27] Initializing ROC-GLM
 #> 
-#> [2022-08-22 13:39:19] Host: Received scores of negative response
-#> [2022-08-22 13:39:19] Receiving negative scores
-#> [2022-08-22 13:39:19] Host: Pushing pooled scores
-#> [2022-08-22 13:39:20] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-08-22 13:39:21] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-08-22 13:39:22] Deviance of iter1=137.2431
-#> [2022-08-22 13:39:23] Deviance of iter2=121.5994
-#> [2022-08-22 13:39:24] Deviance of iter3=147.7237
-#> [2022-08-22 13:39:25] Deviance of iter4=140.4008
-#> [2022-08-22 13:39:26] Deviance of iter5=129.2244
-#> [2022-08-22 13:39:26] Deviance of iter6=123.9979
-#> [2022-08-22 13:39:27] Deviance of iter7=123.1971
-#> [2022-08-22 13:39:28] Deviance of iter8=124.1615
-#> [2022-08-22 13:39:29] Deviance of iter9=124.5356
-#> [2022-08-22 13:39:30] Deviance of iter10=124.5503
-#> [2022-08-22 13:39:30] Deviance of iter11=124.5504
-#> [2022-08-22 13:39:31] Deviance of iter12=124.5504
-#> [2022-08-22 13:39:31] Host: Finished calculating ROC-GLM
-#> [2022-08-22 13:39:31] Host: Cleaning data on server
-#> [2022-08-22 13:39:32] Host: Calculating AUC and CI
-#> [2022-08-22 13:39:39] Finished!
+#> [2022-08-27 13:20:27] Host: Received scores of negative response
+#> [2022-08-27 13:20:27] Receiving negative scores
+#> [2022-08-27 13:20:29] Host: Pushing pooled scores
+#> [2022-08-27 13:20:30] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-08-27 13:20:31] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-08-27 13:20:33] Deviance of iter1=137.2431
+#> [2022-08-27 13:20:34] Deviance of iter2=121.5994
+#> [2022-08-27 13:20:35] Deviance of iter3=147.7237
+#> [2022-08-27 13:20:37] Deviance of iter4=140.4008
+#> [2022-08-27 13:20:38] Deviance of iter5=129.2244
+#> [2022-08-27 13:20:39] Deviance of iter6=123.9979
+#> [2022-08-27 13:20:41] Deviance of iter7=123.1971
+#> [2022-08-27 13:20:42] Deviance of iter8=124.1615
[…]
