+++ paper/paper.bib
@@ -71,3 +71,12 @@
   publisher={JSTOR},
   doi={10.2307/2531595}
 }
+
+@Article{pROC,
+  title = {pROC: an open-source package for R and S+ to analyze and compare ROC curves},
+  author = {Xavier Robin and Natacha Turck and Alexandre Hainard and Natalia Tiberti and Frédérique Lisacek and Jean-Charles Sanchez and Markus Müller},
+  year = {2011},
+  journal = {BMC Bioinformatics},
+  volume = {12},
+  pages = {77},
+}
