### 2022-09-07 #4545 @brunomontezano
@schalkdaniel Yes, the CONTRIBUTING file is enough. Thanks for pointing me out. And also the state of the field that you added is pretty good.
