+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-08-29 13:35:59] L2 sensitivity is: 0.1281
+#> [2022-09-12 13:42:55] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-08-29 13:35:59] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-09-12 13:42:56] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-08-29 13:35:59] Initializing ROC-GLM
+#> [2022-09-12 13:42:56] Initializing ROC-GLM
 #> 
-#> [2022-08-29 13:35:59] Host: Received scores of negative response
-#> [2022-08-29 13:35:59] Receiving negative scores
-#> [2022-08-29 13:36:00] Host: Pushing pooled scores
-#> [2022-08-29 13:36:01] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-08-29 13:36:02] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-08-29 13:36:03] Deviance of iter1=137.2431
-#> [2022-08-29 13:36:04] Deviance of iter2=121.5994
-#> [2022-08-29 13:36:05] Deviance of iter3=147.7237
-#> [2022-08-29 13:36:05] Deviance of iter4=140.4008
-#> [2022-08-29 13:36:06] Deviance of iter5=129.2244
-#> [2022-08-29 13:36:07] Deviance of iter6=123.9979
-#> [2022-08-29 13:36:08] Deviance of iter7=123.1971
-#> [2022-08-29 13:36:09] Deviance of iter8=124.1615
-#> [2022-08-29 13:36:09] Deviance of iter9=124.5356
-#> [2022-08-29 13:36:10] Deviance of iter10=124.5503
-#> [2022-08-29 13:36:11] Deviance of iter11=124.5504
-#> [2022-08-29 13:36:12] Deviance of iter12=124.5504
-#> [2022-08-29 13:36:12] Host: Finished calculating ROC-GLM
-#> [2022-08-29 13:36:12] Host: Cleaning data on server
-#> [2022-08-29 13:36:13] Host: Calculating AUC and CI
-#> [2022-08-29 13:36:20] Finished!
+#> [2022-09-12 13:42:56] Host: Received scores of negative response
+#> [2022-09-12 13:42:56] Receiving negative scores
+#> [2022-09-12 13:42:57] Host: Pushing pooled scores
+#> [2022-09-12 13:42:59] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-09-12 13:43:00] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-09-12 13:43:01] Deviance of iter1=137.2431
+#> [2022-09-12 13:43:03] Deviance of iter2=121.5994
+#> [2022-09-12 13:43:04] Deviance of iter3=147.7237
+#> [2022-09-12 13:43:05] Deviance of iter4=140.4008
+#> [2022-09-12 13:43:06] Deviance of iter5=129.2244
+#> [2022-09-12 13:43:08] Deviance of iter6=123.9979
+#> [2022-09-12 13:43:09] Deviance of iter7=123.1971
+#> [2022-09-12 13:43:10] Deviance of iter8=124.1615
[…]
