### 2022-09-19 #4545 @csoneson
👋🏻 @joao-lima, @dghoshal-lbl - would one of you be available and interested in reviewing this submission for JOSS?

> dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD

Thanks in advance!
