### 2022-09-24 #4545 @csoneson
👋🏻 @mschubert, @gvegayon - would one of you be available and interested in reviewing this submission for JOSS?

dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD

Thanks in advance!
