### 2022-09-27 #4545 @patRyserWelch8
Good morning,

I must apologise for the delay. It has been some quite busy times.

I reviewed and sent you my finding by email last week (i.e., 20/09/2022). I am more than happy to copy it again in this email.

Best wishes,

Pat

General checks

  *   Repository: Is the source code for this software available at the https://github.com/difuture-lmu/dsBinVal<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fdifuture-lmu%2FdsBinVal&data=05%7C01%7Cp.ryser-welch%40open.ac.uk%7C7ab4361f2bbb41ff642908da628dfd35%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C637930659219459695%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=vUoxKStTWmgf%2BUdxIZQIt87e%2FP9f%2BZ3BxbwS2GqtLNU%3D&reserved=0>?  - Yes
  *   License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fopensource.org%2Flicenses%2Falphabetical&data=05%7C01%7Cp.ryser-welch%40open.ac.uk%7C7ab4361f2bbb41ff642908da628dfd35%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C637930659219459695%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=VeGjNoDwL%2F3TITwrfqAAbnYPX64dKwBgV84Gm6q4QbU%3D&reserved=0> software license? yes
  *   Contribution and authorship: Has the submitting author ***@***.***<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fschalkdaniel&data=05%7C01%7Cp.ryser-welch%40open.ac.uk%7C7ab4361f2bbb41ff642908da628dfd35%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C637930659219459695%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=UTUH9qpp4xQScfRUc9WtNgRhwOT43E6Mj8t6XaZDOrI%3D&reserved=0>) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? yes
  *   Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fjoss.readthedocs.io%2Fen%2Flatest%2Fsubmitting.html%23substantial-scholarly-effort&data=05%7C01%7Cp.ryser-welch%40open.ac.uk%7C7ab4361f2bbb41ff642908da628dfd35%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C637930659219459695%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=AzFtOyVwNg9L%2FevOwnguYCDtlrKbJarYgShNodu5eYE%3D&reserved=0>  yes

Functionality

  *   Installation: Does installation proceed as outlined in the documentation?  yes, however, it should specify the version of datashield compliancy.
  *   Functionality: Have the functional claims of the software been confirmed? Yes, some changes need to be implement thought.  Some output of the ROC curves and numerical outcomes are too disclosive. Data can be inferred and reconstructed. Those needs to be looked into.
  *   Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) Yes.

Documentation

  *   A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is? Not clearly. Some more information and explanation in relation to DataSHIELD disclosure controls need to be implemented.  Also no mention of any DataSHIELD data servers code. Another repo needs to be reviewed.
  *   Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. Not yet. it is an issue.
  *   Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). It is challenging without specifying the data onto the DataSHIELD server and Opal version. This is missing.
  *   Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? Not yet.
  *   Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? Some are far too basic. It needs to be inline with DataSHIELD strong testing strategy.  https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/658505761/Testing
  *
  *   Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support - NONE

Software paper

  *   Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? not to my knowledge
  *   A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work? none yet. It needs to state more closely DataSHIELD and how it needs to integrated within the software.
  *   State of the field: Do the authors describe how this software compares to other commonly-used packages? not yet clearly.
  *   Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? It is suitable.
  *   References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fpandoc.org%2FMANUAL.html%23extension-citations&data=05%7C01%7Cp.ryser-welch%40open.ac.uk%7C7ab4361f2bbb41ff642908da628dfd35%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C637930659219459695%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=qcy218zUqYblJzIgy3%2FRCmPCtT8%2F%2BLXjuItb8%2Fe4BfI%3D&reserved=0>? None yet.

Dr Pat Ryser-Welch

Associate lecturer  - T192, T193, T914, TM351,TM352,TM356, TM356, TM470

Referred as: Her/She

Let's remember:

Start by doing what's necessary; then do what's possible; and suddenly you are doing the impossible.

 Francis of Assisi

Mankind remains the most extraordinary computer of all....

John F. Kennedy

It does not matter how slowly you go as long as you do not stop.

Confucius

Read me : https://interestingcomputing.com
