### 2022-10-11 #4545 @patRyserWelch8
Dear Charlotte,

I am really sorry that I am not understanding where exactly, I need to tick those boxes.  I must be rather unintelligent as I cannot appear to find them.  So please send me direct link to these boxes that appear to be causing us both some troubles.

best wishes,

Dr Pat Ryser-Welch

Associate lecturer  - T192, T193, T914, TM351,TM352,TM356, TM356, TM470

Referred as: Her/She

Let's remember:

Start by doing what's necessary; then do what's possible; and suddenly you are doing the impossible.

 Francis of Assisi

Mankind remains the most extraordinary computer of all....

John F. Kennedy

It does not matter how slowly you go as long as you do not stop.

Confucius

Read me : https://interestingcomputing.com
