### 2022-10-11 #4545 @csoneson
Hi @patRyserWelch8 - here's the process:

- Go to https://github.com/openjournals/joss-reviews/issues/4545
- Make a new post where you just type `@editorialbot generate my checklist`
- This will autogenerate a new post in the GitHub issue, with your checklist. You can then edit that post by clicking in the checkboxes.

Thanks!
