### 2022-10-11 #4545 @csoneson
@patRyserWelch8 - I don't think this quite worked out (I don't think it works to reply to the email - if I understand correctly that's what you did above). If you go to this link in your browser: https://github.com/openjournals/joss-reviews/issues/4545, in the bottom of the page you should see something like this (with your picture, not mine):

![Screenshot 2022-10-11 at 11 21 18](https://user-images.githubusercontent.com/5628952/195052094-146fd437-e25e-4687-8576-22e7069f1cf5.png)

In this box (where it says 'Leave a comment' you can type `@editorialbot generate my checklist` and press the 'Comment' button. Could you let me know if you don't see this? (You need to be logged in to your GitHub account).
