### 2022-10-17 #4545 @patRyserWelch8
Hi,

I think I have followed the instructions and copied the content as requested. If it does not work, then I will let you sort it out in the background.

Best wishes,

Dr Pat Ryser-Welch

Associate lecturer  - T192, T193, T914, TM351,TM352,TM356, TM356, TM470

Referred as: Her/She

Let's remember:

Start by doing what's necessary; then do what's possible; and suddenly you are doing the impossible.

 Francis of Assisi

Mankind remains the most extraordinary computer of all....

John F. Kennedy

It does not matter how slowly you go as long as you do not stop.

Confucius

Read me : https://interestingcomputing.com
