### 2022-10-21 #4545 @csoneson
Hi @patRyserWelch8 - sorry for the delay, I was trying to figure out what I can do on my side. Unfortunately I can't generate the checklist for you (as you would not be able to edit it in that case). The best I can do is to paste the checklist into your post above, but that's not great as it can not be formatted properly (GitHub doesn't support markdown formatting in posts done via email replies). If generating the checklist doesn't work, is there some way you can indicate to us which of the items in the list you are satisfied with (and then I can generate the post with the checklist and tick those boxes for you)?

@schalkdaniel - in any case, please see the comments left in the review above (https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535), and feel free to let us know if you have questions.
