+++ README.Rmd
@@ -41,7 +41,7 @@ opalr::opal.logout(opal)
 
 # ROC-GLM and Calibration for DataSHIELD
 
-The package provides functionality to conduct and visualize ROC analysis and calibration on decentralized data. The basis is the DataSHIELD](https://www.datashield.org/) infrastructure for distributed computing. This package provides the calculation of the [__ROC-GLM__](https://www.jstor.org/stable/2676973?seq=1) with [__AUC confidence intervals__](https://www.jstor.org/stable/2531595?seq=1) as well as calibration curves and the brier score. In order to calculate the ROC-GLM or assess calibration it is necessary to push models and predict them at the servers which is also provided by this package. Note that DataSHIELD uses an option `datashield.privacyLevel` to indicate the minimal amount of numbers required to be allowed to share an aggregated value of these numbers. Instead of setting the option, we directly retrieve the privacy level from the [`DESCRIPTION`](https://github.com/difuture-lmu/dsBinVal/blob/master/DESCRIPTION) file each time a function calls for it. This options is set to 5 by default. The methodology is explained in detail [here](https://arxiv.org/abs/2203.10828).
+The package provides functionality to conduct and visualize ROC analysis and calibration on decentralized data. The basis is the [DataSHIELD](https://www.datashield.org/) infrastructure for distributed computing. This package provides the calculation of the [__ROC-GLM__](https://www.jstor.org/stable/2676973?seq=1) with [__AUC confidence intervals__](https://www.jstor.org/stable/2531595?seq=1) as well as calibration curves and the brier score. In order to calculate the ROC-GLM or assess calibration it is necessary to push models and predict them at the servers which is also provided by this package. Note that DataSHIELD uses an option `datashield.privacyLevel` to indicate the minimal amount of numbers required to be allowed to share an aggregated value of these numbers. Instead of setting the option, we directly retrieve the privacy level from the [`DESCRIPTION`](https://github.com/difuture-lmu/dsBinVal/blob/master/DESCRIPTION) file each time a function calls for it. This options is set to 5 by default. The methodological base of th epackage is explained in detail [here](https://arxiv.org/abs/2203.10828).
 
 ## Installation
 
