### 2022-10-25 difuture-lmu/dsBinVal#5 @schalkdaniel
[JOSS REVIEW] Example usage 
From the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535):

> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems):
>> It is challenging without specifying the data onto the DataSHIELD server and Opal version. This is missing.
