### 2022-10-25 difuture-lmu/dsBinVal#6 @schalkdaniel
[JOSS Review] Functionality documentation
From the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535):

> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>> Not yet.
