### 2022-10-25 difuture-lmu/dsBinVal#7 @schalkdaniel
[JOSS Review] Automated tests
From the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535):

> Are there automated tests or manual steps described so that the functionality of the software can be verified?
>> Some are far too basic. It needs to be inline with DataSHIELD strong testing strategy.  https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/658505761/Testing
