### 2022-10-25 difuture-lmu/dsBinVal#8 @schalkdaniel
[JOSS Review] Statement of need
From the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535):

> Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>> Not clearly. Some more information and explanation in relation to DataSHIELD disclosure controls need to be implemented.  Also no mention of any DataSHIELD data servers code. Another repo needs to be reviewed.
