+++ README.Rmd
@@ -58,6 +58,24 @@ if (! check2)
   stop("[", Sys.time(), "] Was not able to publish methods of dsBinVal!")
 
 opalr::opal.logout(opal)
+
+# Build model for the example, therefore download the CNSIM data sets from:
+# https://github.com/datashield/DSLite/tree/master/data
+
+if (FALSE) {
+  dpath = "~/Downloads"
+  dnames = paste0(dpath, "/CNSIM", seq_len(3), ".rda")
+  dLoader = function(n) {
+    load(n)
+    dn = ls()
+    dn = dn[grep("CNSIM", dn)]
+    return(get(dn))
+  }
+  CNSIM = na.omit(do.call(rbind, lapply(dnames, dLoader)))
+  mod = glm(DIS_DIAB ~ ., data = CNSIM, family = binomial())
+  save(mod, file = here::here("Readme_files/mod.rda"))
+}
+
 ```
 [![R-CMD-check](https://github.com/difuture-lmu/dsBinVal/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/difuture-lmu/dsBinVal/actions/workflows/R-CMD-check.yaml) [![License: LGPL v3](https://img.shields.io/badge/License-LGPL%20v3-blue.svg)](https://www.gnu.org/licenses/lgpl-3.0) [![codecov](https://codecov.io/gh/difuture-lmu/dsBinVal/branch/main/graph/badge.svg?token=E8AZRM6XJX)](https://codecov.io/gh/difuture-lmu/dsBinVal)
 
@@ -128,39 +146,44 @@ builder$append(
   server   = "ds1",
   url      = surl,
   user     = username,
-  password = password
+  password = password,
+  table    = "CNSIM.CNSIM1"
 )
 builder$append(
   server   = "ds2",
   url      = surl,
   user     = username,
-  password = password
+  password = password,
+  table    = "CNSIM.CNSIM2"
+)
+builder$append(
+  server   = "ds3",
+  url      = surl,
+  user     = username,
+  password = password,
+  table    = "CNSIM.CNSIM3"
 )
 
 connections = datashield.login(logins = builder$build(), assign = TRUE)
 ```
 
-#### Assign iris and validation vector at DataSHIELD (just for testing)
-
-```{r}
-datashield.assign(connections, "iris", quote(iris))
-vcall = paste0("quote(c(", paste(rep(c(1, 0), times = c(50, 100)), collapse = ", "), "))")
-datashield.assign(connections, "y", eval(parse(text = vcall)))
-```
-
 #### Load test model, push to DataSHIELD, and calculate predictions
 
 ```{r}
-# Model predicts if species of iris is setosa or not.
-iris$y = ifelse(iris$Species == "setosa", 1, 0)
-mod = glm(y ~ Sepal.Length, data = iris, family = binomial())
+# Load the model fitted locally on CNSIM:
+load(here::here("Readme_files/mod.rda"))
+# Model was calculated by:
+#> glm(DIS_DIAB ~ ., data = CNSIM, family = binomial())
 
-# Push the model to the DataSHIELD servers using `dsPredictBase`:
+# Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
 
-# Calculate scores and save at the servers using `dsPredictBase`:
+# Create a clean data set without NAs:
+ds.completeCases("D", newobj = "D_complete")
+
+# Calculate scores and save at the servers:
 pfun =  "predict(mod, newdata = D, type = 'response')"
-predictModel(connections, mod, "pred", "iris", predict_fun = pfun)
+predictModel(connections, mod, "pred", "D_complete", predict_fun = pfun)
 
[…]
