+++ README.md
@@ -49,10 +49,10 @@ server and the analysts machine.
 
 The two options are to use the Opal API:
 
-  - Log into Opal ans switch to the `Administration/DataSHIELD/` tab
-  - Click the `Add DataSHIELD package` button
-  - Select `GitHub` as source, and use `difuture-lmu` as user,
-    `dsBinVal` as name, and `main` as Git reference.
+- Log into Opal ans switch to the `Administration/DataSHIELD/` tab
+- Click the `Add DataSHIELD package` button
+- Select `GitHub` as source, and use `difuture-lmu` as user, `dsBinVal`
+  as name, and `main` as Git reference.
 
 The second option is to use the `opalr` package to install `dsBinVal`
 directly from `R`:
@@ -102,48 +102,76 @@ builder$append(
   server   = "ds1",
   url      = surl,
   user     = username,
-  password = password
+  password = password,
+  table    = "CNSIM.CNSIM1"
 )
 builder$append(
   server   = "ds2",
   url      = surl,
   user     = username,
-  password = password
+  password = password,
+  table    = "CNSIM.CNSIM2"
+)
+builder$append(
+  server   = "ds3",
+  url      = surl,
+  user     = username,
+  password = password,
+  table    = "CNSIM.CNSIM3"
 )
 
 connections = datashield.login(logins = builder$build(), assign = TRUE)
 #> 
 #> Logging into the collaborating servers
-```
-
-#### Assign iris and validation vector at DataSHIELD (just for testing)
-
-``` r
-datashield.assign(connections, "iris", quote(iris))
-vcall = paste0("quote(c(", paste(rep(c(1, 0), times = c(50, 100)), collapse = ", "), "))")
-datashield.assign(connections, "y", eval(parse(text = vcall)))
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Login ds1 [==================>---------------------------------------------------------]  25% / 0s  Login ds2 [=====================================>--------------------------------------]  50% / 0s  Login ds3 [========================================================>-------------------]  75% / 1s  Logged in all servers [================================================================] 100% / 1s
+#> 
+#>   No variables have been specified. 
+#>   All the variables in the table 
+#>   (the whole dataset) will be assigned to R!
+#> 
+#> Assigning table data...
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Assigning ds1 (CNSIM.CNSIM1) [=============>-------------------------------------------]  25% / 1s  Assigning ds2 (CNSIM.CNSIM2) [===========================>-----------------------------]  50% / 1s  Assigning ds3 (CNSIM.CNSIM3) [==========================================>--------------]  75% / 1s  Assigned all tables [==================================================================] 100% / 1s
 ```
 
 #### Load test model, push to DataSHIELD, and calculate predictions
 
 ``` r
-# Model predicts if species of iris is setosa or not.
-iris$y = ifelse(iris$Species == "setosa", 1, 0)
-mod = glm(y ~ Sepal.Length, data = iris, family = binomial())
+# Load the model fitted locally on CNSIM:
+load(here::here("Readme_files/mod.rda"))
+# Model was calculated by:
+#> glm(DIS_DIAB ~ ., data = CNSIM, family = binomial())
 
-# Push the model to the DataSHIELD servers using `dsPredictBase`:
+# Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
+#> [2022-10-26 17:11:56] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
[…]
