+++ README.md
@@ -143,7 +143,7 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
-#> [2022-10-26 17:11:56] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
+#> [2022-10-26 17:17:18] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
 #>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (mod <- decodeBinary("580a000000030004020100030500000000055554462d3800000313000000...  Finalizing assignment ds1 (mod <- decodeBinary("580a000000030004020100030500000000055554462d380...  Checking ds2 (mod <- decodeBinary("580a000000030004020100030500000000055554462d3800000313000000...  Finalizing assignment ds2 (mod <- decodeBinary("580a000000030004020100030500000000055554462d380...  Checking ds3 (mod <- decodeBinary("580a000000030004020100030500000000055554462d3800000313000000...  Finalizing assignment ds3 (mod <- decodeBinary("580a000000030004020100030500000000055554462d380...  Assigned expr. (mod <- decodeBinary("580a000000030004020100030500000000055554462d38000003130000...
 
 # Create a clean data set without NAs:
@@ -200,7 +200,7 @@ l2s
 ``` r
 # The response must be encoded as integer/numeric vector:
 ds.asInteger("D_complete$DIS_DIAB", "truth")
-#>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (exists("DIS_DIAB", D_complete)) [----------------------------------------]   0% / 0s  Getting aggregate ds1 (exists("DIS_DIAB", D_complete)) [=======>-----------------------]  25% / 0s  Checking ds2 (exists("DIS_DIAB", D_complete)) [=========>------------------------------]  25% / 0s  Getting aggregate ds2 (exists("DIS_DIAB", D_complete)) [===============>---------------]  50% / 0s  Checking ds3 (exists("DIS_DIAB", D_complete)) [===================>--------------------]  50% / 0s  Getting aggregate ds3 (exists("DIS_DIAB", D_complete)) [======================>--------]  75% / 0s  Aggregated (exists("DIS_DIAB", D_complete)) [==========================================] 100% / 0s
+#>    [-------------------------------------------------------------------------------------]   0% / 0s  Checking ds1 (exists("DIS_DIAB", D_complete)) [----------------------------------------]   0% / 0s  Getting aggregate ds1 (exists("DIS_DIAB", D_complete)) [=======>-----------------------]  25% / 0s  Checking ds2 (exists("DIS_DIAB", D_complete)) [=========>------------------------------]  25% / 0s  Getting aggregate ds2 (exists("DIS_DIAB", D_complete)) [===============>---------------]  50% / 0s  Checking ds3 (exists("DIS_DIAB", D_complete)) [===================>--------------------]  50% / 0s  Getting aggregate ds3 (exists("DIS_DIAB", D_complete)) [======================>--------]  75% / 0s  Aggregated (exists("DIS_DIAB", D_complete)) [==========================================] 100% / 1s
[…]
