+++ R/brier_score.R
@@ -1,4 +1,4 @@
-#' @title Calculate Brier score
+#' @title Calculate Brier score, called by dsBrierScore
 #' @description This function calculates the Brier score for a given 0-1-vector and
 #'   vector of probabilities.
 #' @param truth_name (`character(1L)`) Character containing the name of the vector of 0-1-values
@@ -8,19 +8,33 @@
 #' @author Daniel S.
 #' @export
 brierScore = function(truth_name, prob_name) {
+
+  #############################################################
+  #MODULE 1: CAPTURE THE nfilter SETTINGS
+  thr = dsBase::listDisclosureSettingsDS()
+  nfilter_tab = as.numeric(thr$nfilter.tab)
+  #nfilter_glm = as.numeric(thr$nfilter.glm)
+  #nfilter_subset = as.numeric(thr$nfilter.subset)
+  #nfilter_string = as.numeric(thr$nfilter.string)
+  #############################################################
+
   checkmate::assertCharacter(truth_name, len = 1L, null.ok = FALSE, any.missing = FALSE)
   checkmate::assertCharacter(prob_name, len = 1L, null.ok = FALSE, any.missing = FALSE)
 
-  truth = eval(parse(text = truth_name))
-  prob  = eval(parse(text = prob_name))
+  #truth = eval(parse(text = truth_name))
+  truth = get(truth_name, envir = parent.frame())
+  #prob = eval(parse(text = prob_name))
+  prob = get(prob_name, envir = parent.frame())
 
   ntruth = length(truth)
   checkmate::assertNumeric(prob, len = ntruth, null.ok = FALSE, any.missing = FALSE)
 
   ## Calculate brier score just if there are at least five or more values to ensure privacy:
-  nfilter_privacy = .getPrivacyLevel()
-  if (ntruth < nfilter_privacy)
-    stop("More than ", nfilter_privacy, " observations are required to ensure privacy!")
+
+  # Fallback if `listDisclosureSettingsDS` returns NULL:
+  if (length(nfilter_tab) == 0) nfilter_tab = .getPrivacyLevel()
+  if (ntruth < nfilter_tab)
+    stop("More than ", nfilter_tab, " observations are required to ensure privacy!")
 
   if (is.character(truth))
     truth = as.integer(as.factor(truth))
