### 2022-11-07 39cba9c
add dsBase as remote repository + add seed for tests
