### 2022-11-07 #4545 @schalkdaniel
@csoneson I looked into @patRyserWelch8 s comment and tried to fix everything as desired. I will try to address all points from [here](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1274382535):

__Functionality:__

- _Installation:_  I have added details about the compliancy in the README (see [this PR](https://github.com/difuture-lmu/dsBinVal/pull/11)).
- _Functionality:_ I cannot comprehend at which point disclosive information is leaked. Maybe @patRyserWelch8 can give further explanations here. The methodology behind the software is designed not to communicate sensitive data. I agree that the empirical ROC curve reveals information. But, since an approximation is used and additionally protected by differential privacy, everything should be fine.

__Documentation:__

- _A statement of need:_ This was indeed a good point, and thanks to @patRyserWelch8 for pointing that out. I updated the software and documented it [here](https://github.com/difuture-lmu/dsBinVal/issues/8). Feel free to comment. Another thing is that I did not split the package into a client and server package because of its size, and I do not think that it should be a big issue concerning the usability of the package with DataSHIELD.
- _Installation instructions:_ I would argue that this is solved by using Rs package managing capabilities via specifying all dependencies in the [DESCRIPTION](https://github.com/difuture-lmu/dsBinVal/blob/main/DESCRIPTION). The installation instructions are described in the [README](https://github.com/difuture-lmu/dsBinVal#installation) of the repository.
- _Example usage:_ This was improved according to to @patRyserWelch8 comment. The concerning issue is [here](https://github.com/difuture-lmu/dsBinVal/issues/5).
-  _Functionality documentation:_ I would also argue that the software is documented. Therefore, I used roxygen, the standard for R software. @patRyserWelch8 can maybe give further explanations of what is missing. Still, I created an [issue](https://github.com/difuture-lmu/dsBinVal/issues/6) to discuss it.
- _Automated tests:_ I tried to understand the mentioned [DataSHIELD testing framework](https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/658505761/Testing), but it was not easy for me to understand how to apply it to a package. For testing, I also spent some effort writing tests not just for the local functions but also for the server functions that get tested on an active DataSHIELD instance.
- _Community guidelines:_ Should also be already there (see [CONTRIBUTING](https://github.com/difuture-lmu/dsBinVal/blob/main/CONTRIBUTING.md) and [COC](https://github.com/difuture-lmu/dsBinVal/blob/main/CODE_OF_CONDUCT.md)).

__Software paper:__

- _Summary:_ I would argue that the paper clearly describes the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- _A statement of need:_ A short statement should also be in the paper.
- _State of the field:_ A short statement was added after @brunomontezano [comment](https://github.com/openjournals/joss-reviews/issues/4545#issuecomment-1210673460).
-  _References:_ References should also be included properly.
