+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-11-07 13:46:12] L2 sensitivity is: 0.1281
+#> [2022-11-14 13:46:29] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-11-07 13:46:13] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-11-14 13:46:30] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-11-07 13:46:13] Initializing ROC-GLM
+#> [2022-11-14 13:46:30] Initializing ROC-GLM
 #> 
-#> [2022-11-07 13:46:13] Host: Received scores of negative response
-#> [2022-11-07 13:46:13] Receiving negative scores
-#> [2022-11-07 13:46:15] Host: Pushing pooled scores
-#> [2022-11-07 13:46:16] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-11-07 13:46:18] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-11-07 13:46:19] Deviance of iter1=137.2431
-#> [2022-11-07 13:46:20] Deviance of iter2=121.5994
-#> [2022-11-07 13:46:21] Deviance of iter3=147.7237
-#> [2022-11-07 13:46:22] Deviance of iter4=140.4008
-#> [2022-11-07 13:46:22] Deviance of iter5=129.2244
-#> [2022-11-07 13:46:23] Deviance of iter6=123.9979
-#> [2022-11-07 13:46:24] Deviance of iter7=123.1971
-#> [2022-11-07 13:46:25] Deviance of iter8=124.1615
-#> [2022-11-07 13:46:26] Deviance of iter9=124.5356
-#> [2022-11-07 13:46:27] Deviance of iter10=124.5503
-#> [2022-11-07 13:46:28] Deviance of iter11=124.5504
-#> [2022-11-07 13:46:29] Deviance of iter12=124.5504
-#> [2022-11-07 13:46:29] Host: Finished calculating ROC-GLM
-#> [2022-11-07 13:46:29] Host: Cleaning data on server
-#> [2022-11-07 13:46:30] Host: Calculating AUC and CI
-#> [2022-11-07 13:46:38] Finished!
+#> [2022-11-14 13:46:30] Host: Received scores of negative response
+#> [2022-11-14 13:46:30] Receiving negative scores
+#> [2022-11-14 13:46:32] Host: Pushing pooled scores
+#> [2022-11-14 13:46:34] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-11-14 13:46:35] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-11-14 13:46:37] Deviance of iter1=137.2431
+#> [2022-11-14 13:46:38] Deviance of iter2=121.5994
+#> [2022-11-14 13:46:39] Deviance of iter3=147.7237
+#> [2022-11-14 13:46:41] Deviance of iter4=140.4008
+#> [2022-11-14 13:46:42] Deviance of iter5=129.2244
+#> [2022-11-14 13:46:43] Deviance of iter6=123.9979
+#> [2022-11-14 13:46:45] Deviance of iter7=123.1971
+#> [2022-11-14 13:46:46] Deviance of iter8=124.1615
[…]
