### 2022-11-24 #4545 @csoneson
> @patRyserWelch8 - could you take a look at the revised submission when you have a chance, and let us know what you think. In particular, please let us know if there are any points in the checklist that you still do not feel are satisfied.

👋🏻 @patRyserWelch8 - could you please confirm that you received the message above about the revised JOSS submission of `dsBinVal`, and that you are willing to take a look at it? If you could provide a timeline of when you think you might get to it, that would be super helpful. Thanks!!
