+++ README.md
@@ -170,36 +170,36 @@ l2s
 roc_glm = dsROCGLM(connections, truth_name = "y", pred_name = "pred",
   dat_name = "iris", seed_object = "y")
 #> 
-#> [2022-12-13 08:24:09] L2 sensitivity is: 0.1281
+#> [2022-12-19 13:17:57] L2 sensitivity is: 0.1281
 #> Warning in dsROCGLM(connections, truth_name = "y", pred_name = "pred", dat_name
 #> = "iris", : l2-sensitivity may be too high for good results! Epsilon = 0.5 and
 #> delta = 0.5 is used which may lead to bad results.
 #> 
-#> [2022-12-13 08:24:10] Setting: epsilon = 0.5 and delta = 0.5
+#> [2022-12-19 13:17:58] Setting: epsilon = 0.5 and delta = 0.5
 #> 
-#> [2022-12-13 08:24:10] Initializing ROC-GLM
+#> [2022-12-19 13:17:58] Initializing ROC-GLM
 #> 
-#> [2022-12-13 08:24:10] Host: Received scores of negative response
-#> [2022-12-13 08:24:10] Receiving negative scores
-#> [2022-12-13 08:24:11] Host: Pushing pooled scores
-#> [2022-12-13 08:24:12] Server: Calculating placement values and parts for ROC-GLM
-#> [2022-12-13 08:24:12] Server: Calculating probit regression to obtain ROC-GLM
-#> [2022-12-13 08:24:13] Deviance of iter1=137.2431
-#> [2022-12-13 08:24:14] Deviance of iter2=121.5994
-#> [2022-12-13 08:24:15] Deviance of iter3=147.7237
-#> [2022-12-13 08:24:16] Deviance of iter4=140.4008
-#> [2022-12-13 08:24:16] Deviance of iter5=129.2244
-#> [2022-12-13 08:24:17] Deviance of iter6=123.9979
-#> [2022-12-13 08:24:18] Deviance of iter7=123.1971
-#> [2022-12-13 08:24:18] Deviance of iter8=124.1615
-#> [2022-12-13 08:24:19] Deviance of iter9=124.5356
-#> [2022-12-13 08:24:20] Deviance of iter10=124.5503
-#> [2022-12-13 08:24:21] Deviance of iter11=124.5504
-#> [2022-12-13 08:24:21] Deviance of iter12=124.5504
-#> [2022-12-13 08:24:21] Host: Finished calculating ROC-GLM
-#> [2022-12-13 08:24:21] Host: Cleaning data on server
-#> [2022-12-13 08:24:22] Host: Calculating AUC and CI
-#> [2022-12-13 08:24:29] Finished!
+#> [2022-12-19 13:17:58] Host: Received scores of negative response
+#> [2022-12-19 13:17:58] Receiving negative scores
+#> [2022-12-19 13:17:59] Host: Pushing pooled scores
+#> [2022-12-19 13:18:01] Server: Calculating placement values and parts for ROC-GLM
+#> [2022-12-19 13:18:02] Server: Calculating probit regression to obtain ROC-GLM
+#> [2022-12-19 13:18:03] Deviance of iter1=137.2431
+#> [2022-12-19 13:18:05] Deviance of iter2=121.5994
+#> [2022-12-19 13:18:06] Deviance of iter3=147.7237
+#> [2022-12-19 13:18:07] Deviance of iter4=140.4008
+#> [2022-12-19 13:18:08] Deviance of iter5=129.2244
+#> [2022-12-19 13:18:10] Deviance of iter6=123.9979
+#> [2022-12-19 13:18:11] Deviance of iter7=123.1971
+#> [2022-12-19 13:18:12] Deviance of iter8=124.1615
[…]
