+++ README.md
@@ -17,13 +17,16 @@ confidence intervals**](https://www.jstor.org/stable/2531595?seq=1) as
 well as calibration curves and the brier score. In order to calculate
 the ROC-GLM or assess calibration it is necessary to push models and
 predict them at the servers which is also provided by this package. Note
-that DataSHIELD uses an option `datashield.privacyLevel` to indicate the
-minimal amount of numbers required to be allowed to share an aggregated
-value of these numbers. Instead of setting the option, we directly
-retrieve the privacy level from the
+that DataSHIELD uses [privacy
+filter](https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/714768398/Disclosure+control)
+from DataSHIELD v5 onwards that are also used in this package.
+Additionally, this package uses the old option `datashield.privacyLevel`
+(to indicate the minimal amount of values required to allow sharing an
+aggregation) as fallback. Instead of setting the option, we directly
+retrieve the fallback privacy level from the
 [`DESCRIPTION`](https://github.com/difuture-lmu/dsBinVal/blob/master/DESCRIPTION)
 file each time a function calls for it. This options is set to 5 by
-default. The methodological base of th epackage is explained in detail
+default. The methodological base of the package is explained in detail
 [here](https://arxiv.org/abs/2203.10828).
 
 ## Installation
@@ -49,10 +52,10 @@ server and the analysts machine.
 
 The two options are to use the Opal API:
 
-- Log into Opal ans switch to the `Administration/DataSHIELD/` tab
-- Click the `Add DataSHIELD package` button
-- Select `GitHub` as source, and use `difuture-lmu` as user, `dsBinVal`
-  as name, and `main` as Git reference.
+  - Log into Opal ans switch to the `Administration/DataSHIELD/` tab
+  - Click the `Add DataSHIELD package` button
+  - Select `GitHub` as source, and use `difuture-lmu` as user,
+    `dsBinVal` as name, and `main` as Git reference.
 
 The second option is to use the `opalr` package to install `dsBinVal`
 directly from `R`:
@@ -123,14 +126,12 @@ builder$append(
 connections = datashield.login(logins = builder$build(), assign = TRUE)
 #> 
 #> Logging into the collaborating servers
-#>    [-------------------------------------------------------------------------------------]   0% / 0s  Login ds1 [==================>---------------------------------------------------------]  25% / 0s  Login ds2 [=====================================>--------------------------------------]  50% / 0s  Login ds3 [========================================================>-------------------]  75% / 1s  Logged in all servers [================================================================] 100% / 1s
 #> 
 #>   No variables have been specified. 
 #>   All the variables in the table 
 #>   (the whole dataset) will be assigned to R!
 #> 
 #> Assigning table data...
-#>    [-------------------------------------------------------------------------------------]   0% / 0s  Assigning ds1 (CNSIM.CNSIM1) [=============>-------------------------------------------]  25% / 1s  Assigning ds2 (CNSIM.CNSIM2) [===========================>-----------------------------]  50% / 1s  Assigning ds3 (CNSIM.CNSIM3) [==========================================>--------------]  75% / 1s  Assigned all tables [==================================================================] 100% / 1s
 ```
 
 #### Load test model, push to DataSHIELD, and calculate predictions
@@ -143,15 +144,10 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
[…]
