### 2023-01-14 #4545 @csoneson
Happy New Year everyone - let's try to find another reviewer who would be willing to take a look at this submission 🙂

@ankur-gupta, @AnthonyOfSeattle - would one of you be interested in reviewing this submission for JOSS?

> dsBinVal: Conducting distributed ROC analysis and Calibration using DataSHIELD
