+++ man/checkTruthProb.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_scores.R
+% Please edit documentation in R/helper.R
 \name{checkTruthProb}
 \alias{checkTruthProb}
 \title{Truth and Prediction Checker}
