+++ README.Rmd
@@ -82,7 +82,7 @@ if (FALSE) {
 # ROC-GLM and Calibration for DataSHIELD
 
 
-The package provides functionality to conduct and visualize ROC analysis and calibration on decentralized data. The basis is the [DataSHIELD](https://www.datashield.org/) infrastructure for distributed computing. This package provides the calculation of the [**ROC-GLM**](https://www.jstor.org/stable/2676973?seq=1) with [**AUC confidence intervals**](https://www.jstor.org/stable/2531595?seq=1) as well as calibration curves and the brier score. In order to calculate the ROC-GLM or assess calibration it is necessary to push models and predict them at the servers which is also provided by this package. Note that DataSHIELD uses [privacy filter](https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/714768398/Disclosure+control) from DataSHIELD v5 onwards that are also used in this package. Additionally, this package uses the old option `datashield.privacyLevel` (to indicate the minimal amount of values required to allow sharing an aggregation) as fallback. Instead of setting the option, we directly retrieve the fallback privacy level from the [`DESCRIPTION`](https://github.com/difuture-lmu/dsBinVal/blob/master/DESCRIPTION) file each time a function calls for it. This options is set to 5 by default. The methodological base of the package is explained in detail [here](https://arxiv.org/abs/2203.10828).
+The package provides functionality to conduct and visualize ROC analysis and calibration on decentralized data. The basis is the [DataSHIELD](https://www.datashield.org/) infrastructure for distributed computing. This package provides the calculation of the [**ROC-GLM**](https://www.jstor.org/stable/2676973?seq=1) with [**AUC confidence intervals**](https://www.jstor.org/stable/2531595?seq=1) as well as [**calibration curves**](https://www.geeksforgeeks.org/calibration-curves/) and the [**Brier score**](https://en.wikipedia.org/wiki/Brier_score). In order to calculate the ROC-GLM or assess calibration it is necessary to push models and predict them at the servers which is also provided by this package. Note that DataSHIELD uses [privacy filter](https://data2knowledge.atlassian.net/wiki/spaces/DSDEV/pages/714768398/Disclosure+control) from DataSHIELD v5 onwards that are also used in this package. Additionally, this package uses the old option `datashield.privacyLevel` (to indicate the minimal amount of values required to allow sharing an aggregation) as fallback. Instead of setting the option, we directly retrieve the fallback privacy level from the [`DESCRIPTION`](https://github.com/difuture-lmu/dsBinVal/blob/master/DESCRIPTION) file each time a function calls for it. This options is set to 5 by default. The methodology of the package is explained in detail [here](https://arxiv.org/abs/2203.10828).
 
 ## Installation
 
