### 2023-01-16 86c3049
Update NEWS.md [ci skip]
