+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ##
 
+- **16.01.2023** \
+  Remove `ggplot2::aes_string` and use `ggplot2::aes` instead.
+
 - **12.05.2022** \
   Add method to calculate the confusion matrix and measures like f1 score, fpr, tpr, and so on.
 
