### 2023-01-17 #4545 @AnthonyOfSeattle
@csoneson, I would be happy to help. Given @patRyserWelch8's background in security, I am also interested to hear their elaboration on the disclosive aspects of the algorithm, which could be possibly addressed in a `Maintaining Security` section of the README.

While my background is not in security, I see this software is related to another publication (https://arxiv.org/abs/2203.10828).  Given that, I would be comfortable evaluating the quality of the software with respect to it implementing the functionality described in the citations. However, I would not be comfortable evaluating the general security of the algorithms themselves.

If that is acceptable, I am happy to start reviewing this week.
