### 2023-01-20 difuture-lmu/dsBinVal#12 @AnthonyOfSeattle
Copy editing of software paper
Please address the following copy-writing edits in the paper if they accurately reflect what you want to say.

**Line:** [38](https://github.com/difuture-lmu/dsBinVal/blob/e01c1c42296556e4f7354486a4ed3595ea643311/paper/paper.md?plain=1#L38)
**Original:** the AUC and confidence intervals after DeLong et al. (1988) to conduct hypothesis testing are derived.
**Requested change:** the AUC and confidence intervals are derived to conduct hypothesis testing according to DeLong et al. (1988).

**Line:** [42](https://github.com/difuture-lmu/dsBinVal/blob/e01c1c42296556e4f7354486a4ed3595ea643311/paper/paper.md?plain=1#L42)
**Original 1:** Uncontrolled release of health information may imply personal disadvantages for individuals.
**Requested change 1:** Uncontrolled release of health information may cause personal disadvantages for individuals
**Original 2:** The individual patient needs to be protected that personal features become visible to people not authorized to know them.
**Requested change 2:** The individual patient needs to be protected against personal details becoming visible to people not authorized to know them.

**Line:** [44](https://github.com/difuture-lmu/dsBinVal/blob/e01c1c42296556e4f7354486a4ed3595ea643311/paper/paper.md?plain=1#L44)
**Original 1:** Furthermore, using confidential data comes with administrative burdens and mostly requires a consent about using the data.
**Requested change 1:** Do you mean "requires consent around data usage" ?
**Original 2:** Modern approaches in distributed analysis allow to work on distributed confidential data by providing frameworks that allow retrieval of information without sharing sensitive information.
**Requested change 2:** Modern approaches in distributed analysis allow work on distributed confidential data by providing frameworks that allow retrieval of information without sharing sensitive information.

**Line:** [46](https://github.com/difuture-lmu/dsBinVal/blob/e01c1c42296556e4f7354486a4ed3595ea643311/paper/paper.md?plain=1#L46)
**Original 1:** The framework already provides, among others, techniques for descriptive statistics, basic summary statistics, or basic statistical modeling.
**Requested change 1:**The framework already provides techniques for descriptive statistics, basic summary statistics, and basic statistical modeling.
**Orginal 2:** This model is to be validated using ROC and calibration analysis on patient data distributed across five hospitals using DataSHIELD.
**Requested change 2:** Is this supposed to be past tense? Are you describing an experiment you already performed.
**Original 3:** Distributed ROC analysis and the assessment of calibration is currently not available in DataSHIELD.
**Requested change  3:** This seems out of place. Maybe it should be placed a couple sentences back.
