### 2023-01-20 difuture-lmu/dsBinVal#13 @AnthonyOfSeattle
Address language in paper
The paper states the following:
>These techniques alleviate many of the administrative, ethical and legal requirements on medical research.

It is unclear how distributed analysis alleviates any of these requirements. I would understand if this statement was about how distributed techniques allow researchers to more easily meet these requirements.

I left this comment out of the copy-editing issue since it is more substantive.
