+++ paper/paper.md
@@ -35,17 +35,17 @@ bibliography: paper.bib
 
 # Summary
 
-Our `R` [@rcore] package `dsBinVal` implements the methodology explained by @schalk2022rocglm. It extends the ROC-GLM [@pepe2000interpretation] to distributed data by using techniques of differential privacy [@dwork2006calibrating] and the idea of sharing highly aggregated values only. The package also exports functionality to calculate distributed calibration curves and assess the calibration. Using the package allows us to evaluate a prognostic model based on a binary outcome using the DataSHIELD [@gaye2014datashield] framework. Therefore, the main functionality makes it able to 1) compute the ROC curve using the ROC-GLM from which 2) the AUC and confidence intervals after @delong1988comparing to conduct hypothesis testing are derived. Furthermore, 3) the calibration can be assessed distributively via calibration curves and the Brier score. Visualizing the approximated ROC curve, the AUC with confidence intervals, and the calibration curves is also supported based on [`ggplot2`](https://ggplot2.tidyverse.org/reference/ggplot.html). Examples can be found in the [README](https://github.com/difuture-lmu/dsBinVal) file of the repository.
+Our `R` [@rcore] package `dsBinVal` implements the methodology explained by @schalk2022rocglm. It extends the ROC-GLM [@pepe2000interpretation] to distributed data by using techniques of differential privacy [@dwork2006calibrating] and the idea of sharing highly aggregated values only. The package also exports functionality to calculate distributed calibration curves and assess the calibration. Using the package allows us to evaluate a prognostic model based on a binary outcome using the DataSHIELD [@gaye2014datashield] framework. Therefore, the main functionality makes it able to 1) compute the ROC curve using the ROC-GLM from which 2) the AUC and confidence intervals are derived to conduct hypothesis testing according to @delong1988. Furthermore, 3) the calibration can be assessed distributively via calibration curves and the Brier score. Visualizing the approximated ROC curve, the AUC with confidence intervals, and the calibration curves is also supported based on [`ggplot2`](https://ggplot2.tidyverse.org/reference/ggplot.html). Examples can be found in the [README](https://github.com/difuture-lmu/dsBinVal) file of the repository.
 
 # Statement of need
 
-Privacy protection of patient data plays a major role for a variety of tasks in medical research. Uncontrolled release of health information may imply personal disadvantages for individuals. The individual patient needs to be protected that personal features become visible to people not authorized to know them.
+Privacy protection of patient data plays a major role for a variety of tasks in medical research. Uncontrolled release of health information may cause personal disadvantages for individuals. The individual patient needs to be protected against personal details becoming visible to people not authorized to know them..
 
-In statistics or machine learning, one of these tasks is to gain insights by building statistical or prognostic models. Prognosis on the development of severe health conditions or covariates coding critical health information like genetic susceptibility need to be handled with care. Furthermore, using confidential data comes with administrative burdens and mostly requires a consent about using the data. Additionally, the data can be distributed over multiple sites (e.g. hospitals) which makes their access even more challenging. Modern approaches in distributed analysis allow to work on distributed confidential data by providing frameworks that allow retrieval of information without sharing sensitive information. These techniques alleviate many of the administrative, ethical and legal requirements on medical research.
[…]
