### 2023-01-23 difuture-lmu/dsBinVal#12 @schalkdaniel
@AnthonyOfSeattle, thank you so much for the suggestions. I updated the manuscript as requested (see [7fbd0ec](https://github.com/difuture-lmu/dsBinVal/commit/7fbd0ecf30b4769a7f15f0546543c57f3da9e128)).

Regarding Line [46](https://github.com/difuture-lmu/dsBinVal/blob/e01c1c42296556e4f7354486a4ed3595ea643311/paper/paper.md?plain=1#L46), I changed the sentences as followed:

- **Original 2:** This model is to be validated using ROC and calibration analysis on patient data distributed across five hospitals using DataSHIELD.
- **Requested Change 2:** Is this supposed to be past tense? Are you describing an experiment you already performed.
    - **Update:** "One goal of the multiple sclerosis use-case is to validate that prognostic model using ROC and calibration analysis on patient data distributed across five hospitals using DataSHIELD."

- **Original 3:** Distributed ROC analysis and the assessment of calibration is currently not available in DataSHIELD.
- **Requested change 3:** This seems out of place. Maybe it should be placed a couple sentences back.
    - **Update:** I put it at the end of the next paragraph: "In this package we close the gap between distributed model building and the validation of binary outcomes also on the distributed data. Therefore, our package seamlessly integrates into the DataSHIELD framework, which does not yet provide distributed ROC analysis and calibration assessment."
