+++ paper/paper.md
@@ -39,9 +39,9 @@ Our `R` [@rcore] package `dsBinVal` implements the methodology explained by @sch
 
 # Statement of need
 
-Privacy protection of patient data plays a major role for a variety of tasks in medical research. Uncontrolled release of health information may cause personal disadvantages for individuals. The individual patient needs to be protected against personal details becoming visible to people not authorized to know them..
+Privacy protection of patient data plays a major role for a variety of tasks in medical research. Uncontrolled release of health information may cause personal disadvantages for individuals. The individual patient needs to be protected against personal details becoming visible to people not authorized to know them.
 
-In statistics or machine learning, one of these tasks is to gain insights by building statistical or prognostic models. Prognosis on the development of severe health conditions or covariates coding critical health information like genetic susceptibility need to be handled with care. Furthermore, using confidential data comes with administrative burdens and mostly requires a consent around data usage. Additionally, the data can be distributed over multiple sites (e.g. hospitals) which makes their access even more challenging. Modern approaches in distributed analysis allow work on distributed confidential data by providing frameworks that allow retrieval of information without sharing sensitive information. Since no sensitive information is shared through the use of privacy-preserving and distributed algorithms, their use helps to alleviate administrative, ethical, and legal requirements as researchers do not have access to personal data.
+In statistics or machine learning, one of these tasks is to gain insights by building statistical or prognostic models. Prognosis on the development of severe health conditions or covariates coding critical health information like genetic susceptibility need to be handled with care. Furthermore, using confidential data comes with administrative burdens and mostly requires a consent around data usage. Additionally, the data can be distributed over multiple sites (e.g. hospitals) which makes their access even more challenging. Modern approaches in distributed analysis allow work on distributed confidential data by providing frameworks that allow retrieval of information without sharing sensitive information. Since no sensitive information is shared through the use of privacy-preserving and distributed algorithms, their use helps to alleviate administrative, ethical, and legal requirements in medical research as users do not have access to personal data.
 
 One of these frameworks for privacy protected analysis is DataSHIELD [@gaye2014datashield]. It allows the analysis of data in a non-disclosive setting. The framework already provides techniques for descriptive statistics, basic summary statistics, and basic statistical modeling. Within a multiple sclerosis use-case to enhance patient medication in the DIFUTURE consortium of the German Medical Informatics Initiative [@prasser2018difuture], a prognostic model was developed on individual patient data. One goal of the multiple sclerosis use-case is to validate that prognostic model using ROC and calibration analysis on patient data distributed across five hospitals using DataSHIELD.
 
