### 2023-01-23 difuture-lmu/dsBinVal#13 @schalkdaniel
Thanks, this is a good point. I tried to make it more clear by adjusting the text:

- **Original:** "These techniques alleviate many of the administrative, ethical and legal requirements on medical research."
- **Update:** "Since no sensitive information is shared through the use of privacy-preserving and distributed algorithms, their use helps to alleviate administrative, ethical, and legal requirements in medical research as users do not have access to personal data."
