### 2023-01-25 difuture-lmu/dsBinVal#14 @AnthonyOfSeattle
Broken link to example
https://github.com/difuture-lmu/dsBinVal/blob/60d9946a81d120c706806b37b30d0ad8b95098c7/README.md?plain=1#L83
