+++ README.md
@@ -54,10 +54,10 @@ server and the analysts machine.
 
 The two options are to use the Opal API:
 
-  - Log into Opal ans switch to the `Administration/DataSHIELD/` tab
-  - Click the `Add DataSHIELD package` button
-  - Select `GitHub` as source, and use `difuture-lmu` as user,
-    `dsBinVal` as name, and `main` as Git reference.
+- Log into Opal ans switch to the `Administration/DataSHIELD/` tab
+- Click the `Add DataSHIELD package` button
+- Select `GitHub` as source, and use `difuture-lmu` as user, `dsBinVal`
+  as name, and `main` as Git reference.
 
 The second option is to use the `opalr` package to install `dsBinVal`
 directly from `R`:
@@ -80,7 +80,7 @@ opalr::opal.logout(opal)
 ## Usage
 
 A more sophisticated example is available
-[here](github.com/difuture-lmu/datashield-roc-glm-demo).
+[here](https://github.com/difuture-lmu/datashield-roc-glm-demo).
 
 ``` r
 library(DSI)
@@ -146,7 +146,7 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
-#> [2023-01-23 13:32:28] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
+#> [2023-01-25 08:03:51] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
 
 # Create a clean data set without NAs:
 ds.completeCases("D", newobj = "D_complete")
@@ -202,27 +202,27 @@ ds.asInteger("D_complete$DIS_DIAB", "truth")
 roc_glm = dsROCGLM(connections, truth_name = "truth", pred_name = "pred",
   dat_name = "D_complete", seed_object = "pred")
 #> 
-#> [2023-01-23 13:34:41] L2 sensitivity is: 0.0015
+#> [2023-01-25 08:05:57] L2 sensitivity is: 0.0015
 #> 
-#> [2023-01-23 13:34:43] Setting: epsilon = 0.2 and delta = 0.1
+#> [2023-01-25 08:05:58] Setting: epsilon = 0.2 and delta = 0.1
 #> 
-#> [2023-01-23 13:34:43] Initializing ROC-GLM
+#> [2023-01-25 08:05:58] Initializing ROC-GLM
 #> 
-#> [2023-01-23 13:34:43] Host: Received scores of negative response
-#> [2023-01-23 13:34:43] Receiving negative scores
-#> [2023-01-23 13:34:44] Host: Pushing pooled scores
-#> [2023-01-23 13:34:46] Server: Calculating placement values and parts for ROC-GLM
-#> [2023-01-23 13:34:48] Server: Calculating probit regression to obtain ROC-GLM
-#> [2023-01-23 13:34:49] Deviance of iter1=38.8162
-#> [2023-01-23 13:34:50] Deviance of iter2=48.9408
-#> [2023-01-23 13:34:52] Deviance of iter3=52.5077
-#> [2023-01-23 13:34:53] Deviance of iter4=52.5684
-#> [2023-01-23 13:34:54] Deviance of iter5=52.5684
-#> [2023-01-23 13:34:56] Deviance of iter6=52.5684
-#> [2023-01-23 13:34:56] Host: Finished calculating ROC-GLM
-#> [2023-01-23 13:34:56] Host: Cleaning data on server
-#> [2023-01-23 13:34:57] Host: Calculating AUC and CI
-#> [2023-01-23 13:35:08] Finished!
+#> [2023-01-25 08:05:58] Host: Received scores of negative response
[…]
