### 2023-01-27 difuture-lmu/dsBinVal#12 @AnthonyOfSeattle
Good changes. Thank you for looking at these. There are a couple more minor typographical errors which I can list if you want them. However, I think this issue is complete.
