### 2023-01-30 difuture-lmu/dsBinVal#15 @AnthonyOfSeattle
ggplot2 is not a required package for installd
This is a small nit and probably won't be an issue for most people since most people will already have this installed. However, since plot the ROC curve is fairly standard, and is an advertised function of the package, ggplot2 should be included during automatic install.
