### 2023-01-30 difuture-lmu/dsBinVal#13 @AnthonyOfSeattle
But I believe what I am saying is that the `administrative, ethical, and legal requirements` are still there. They aren't `alleviated`. Do you mean `their use helps to address...` or `their use helps researchers to meet...`?
