### 2023-01-30 difuture-lmu/dsBinVal#13 @schalkdaniel
Now I get it, thanks for pointing it out again. I changed `alleviated` to `meet`, see [4336f13](https://github.com/difuture-lmu/dsBinVal/commit/4336f136887765acbee3e02ca1b06d4c39d8cc7b)
