### 2023-02-06 difuture-lmu/dsBinVal#13 @AnthonyOfSeattle
Thank you for this change. I think it reads well.
