### 2023-02-06 difuture-lmu/dsBinVal#16 @AnthonyOfSeattle
Package help for tutorial
I see a couple packages in the tutorial that did not make it into the depends or imports section of the description:

https://github.com/difuture-lmu/dsBinVal/blob/b87b01d9ec96f2b09f5dc0ffade0386c8b685555/README.md?plain=1#L85-L95

I'd say, place these in the description in the appropriate place. If for some reason they really shouldn't be in Imports, please put a brief line around where I linked which tells users they need to install these.
