+++ README.md
@@ -146,7 +146,7 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
-#> [2023-01-30 13:24:25] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
+#> [2023-02-06 13:32:02] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
 
 # Create a clean data set without NAs:
 ds.completeCases("D", newobj = "D_complete")
@@ -202,27 +202,27 @@ ds.asInteger("D_complete$DIS_DIAB", "truth")
 roc_glm = dsROCGLM(connections, truth_name = "truth", pred_name = "pred",
   dat_name = "D_complete", seed_object = "pred")
 #> 
-#> [2023-01-30 13:26:39] L2 sensitivity is: 0.0015
+#> [2023-02-06 13:34:14] L2 sensitivity is: 0.0015
 #> 
-#> [2023-01-30 13:26:41] Setting: epsilon = 0.2 and delta = 0.1
+#> [2023-02-06 13:34:16] Setting: epsilon = 0.2 and delta = 0.1
 #> 
-#> [2023-01-30 13:26:41] Initializing ROC-GLM
+#> [2023-02-06 13:34:16] Initializing ROC-GLM
 #> 
-#> [2023-01-30 13:26:41] Host: Received scores of negative response
-#> [2023-01-30 13:26:41] Receiving negative scores
-#> [2023-01-30 13:26:42] Host: Pushing pooled scores
-#> [2023-01-30 13:26:44] Server: Calculating placement values and parts for ROC-GLM
-#> [2023-01-30 13:26:46] Server: Calculating probit regression to obtain ROC-GLM
-#> [2023-01-30 13:26:47] Deviance of iter1=38.8162
-#> [2023-01-30 13:26:48] Deviance of iter2=48.9408
-#> [2023-01-30 13:26:50] Deviance of iter3=52.5077
-#> [2023-01-30 13:26:51] Deviance of iter4=52.5684
-#> [2023-01-30 13:26:52] Deviance of iter5=52.5684
-#> [2023-01-30 13:26:54] Deviance of iter6=52.5684
-#> [2023-01-30 13:26:54] Host: Finished calculating ROC-GLM
-#> [2023-01-30 13:26:54] Host: Cleaning data on server
-#> [2023-01-30 13:26:55] Host: Calculating AUC and CI
-#> [2023-01-30 13:27:06] Finished!
+#> [2023-02-06 13:34:16] Host: Received scores of negative response
+#> [2023-02-06 13:34:16] Receiving negative scores
+#> [2023-02-06 13:34:17] Host: Pushing pooled scores
+#> [2023-02-06 13:34:20] Server: Calculating placement values and parts for ROC-GLM
+#> [2023-02-06 13:34:21] Server: Calculating probit regression to obtain ROC-GLM
+#> [2023-02-06 13:34:22] Deviance of iter1=38.8162
+#> [2023-02-06 13:34:24] Deviance of iter2=48.9408
+#> [2023-02-06 13:34:25] Deviance of iter3=52.5077
+#> [2023-02-06 13:34:26] Deviance of iter4=52.5684
+#> [2023-02-06 13:34:28] Deviance of iter5=52.5684
+#> [2023-02-06 13:34:29] Deviance of iter6=52.5684
+#> [2023-02-06 13:34:29] Host: Finished calculating ROC-GLM
+#> [2023-02-06 13:34:29] Host: Cleaning data on server
+#> [2023-02-06 13:34:30] Host: Calculating AUC and CI
[…]
