+++ README.Rmd
@@ -127,10 +127,6 @@ opalr::opal.logout(opal)
 A more sophisticated example is available [here](github.com/difuture-lmu/datashield-roc-glm-demo).
 
 ```{r}
-library(DSI)
-library(DSOpal)
-library(dsBaseClient)
-
 library(dsBinVal)
 ```
 
