+++ DESCRIPTION
@@ -27,7 +27,8 @@ Imports:
     dsBase,
     checkmate,
     stringr,
-    digest
+    digest,
+    dsBaseClient
 Suggests:
     DSOpal,
     ggplot2,
@@ -54,6 +55,7 @@ AssignMethods:
     assignPredictModel,
     removeMissings
 Remotes:
+    datashield/dsBaseClient
     datashield/dsBase
 Options: datashield.privacyLevel=5
 RoxygenNote: 7.1.1
