+++ DESCRIPTION
@@ -26,8 +26,7 @@ Imports:
     DSI,
     dsBase,
     digest,
-    dsBaseClient
-Requires:
+    dsBaseClient,
     checkmate,
     stringr,
     ggplot2
