### 2023-02-09 difuture-lmu/dsBinVal#16 @schalkdaniel
They are now all in the `DESCRIPTION`:  https://github.com/difuture-lmu/dsBinVal/blob/8af612616447f5edaa359eda4a72ae7ad7f8f0c8/DESCRIPTION#L23-L36
