### 2023-02-10 difuture-lmu/dsBinVal#18 @AnthonyOfSeattle
Improve technical details section of the paper
I believe the technical details of the software paper need to be improved, and may benefit from incoporating some of the functionality descriptions in Section 6 of https://arxiv.org/abs/2203.10828 .

This does not need to go in depth, since the theory is laid out in the companion paper. It should really just focus on a short description of what helper functions are provided and what occurs in major functions.

I think a few points could go a long way.
- Provide a brief mention of the statistical functions which are included and how they can be used on a given project, and why.
- Mention that data is returned from the server twice in the roc_glm function calls and note the DP step taken to ensure this is safe.
- Mention that individual functions attempt to prevent accidental disclosures when data is not sufficient.
- Provide a description of the tests done in the package and how they demostrate the safety of the methods.
