### 2023-02-10 #4545 @AnthonyOfSeattle
The authors provide a mature R package for calculating ROC curves in a distributed and safe fashion across data centers. Their framework of choice is the Datashield secure computing software, which appears to be a trusted ecosystem for the intended audience of the tool. In order to perform calculations, the authors return model performance data from individual secure servers. However, this data is protected by a differential privacy approach, which is sufficient detailed in a companion paper. The level of privacy is protected by a function call which derives privacy parameters from the data, which limits the amount of knowledge that the user needs to have before using the ROC GLM functions. The choice in privacy parameters is backed up by a simulation experiment in the companion data. The final AUC is calculated in a non-disclosive fashion by aggregating results from distributed GLM approximations of the ROC. A method to calculate calibration of the distributed component datasets is also provided.

The tutorial was fairly comprehensive and worked as written once the package and dependencies were installed. I particularly enjoyed the amount of helper and statistical functions included, and I expect these will find a fair amount of use in the toolbox of users. The repository also includes automated tests whose results are displayed transparently in the README, which I found to be a nice touch.

I believe that the software paper needs some revisions to include more description of the functionality of the package, where data is disclosed and how it is protected, and the relationship of choices made in the software to the companion paper. I believe this can be done fairly succinctly by referring most technical details to the companion paper. The work is part of the following issue:

https://github.com/difuture-lmu/dsBinVal/issues/18

During the review process, I detailed a few more minor issues. Many of these are already completed.

https://github.com/difuture-lmu/dsBinVal/issues/12
https://github.com/difuture-lmu/dsBinVal/issues/13
https://github.com/difuture-lmu/dsBinVal/issues/14
https://github.com/difuture-lmu/dsBinVal/issues/15
https://github.com/difuture-lmu/dsBinVal/issues/16
