+++ README.md
@@ -138,7 +138,7 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
-#> [2023-02-09 14:17:23] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
+#> [2023-02-10 14:39:27] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
 
 # Create a clean data set without NAs:
 ds.completeCases("D", newobj = "D_complete")
@@ -194,27 +194,27 @@ ds.asInteger("D_complete$DIS_DIAB", "truth")
 roc_glm = dsROCGLM(connections, truth_name = "truth", pred_name = "pred",
   dat_name = "D_complete", seed_object = "pred")
 #> 
-#> [2023-02-09 14:19:38] L2 sensitivity is: 0.0015
+#> [2023-02-10 14:41:25] L2 sensitivity is: 0.0015
 #> 
-#> [2023-02-09 14:19:39] Setting: epsilon = 0.2 and delta = 0.1
+#> [2023-02-10 14:41:26] Setting: epsilon = 0.2 and delta = 0.1
 #> 
-#> [2023-02-09 14:19:39] Initializing ROC-GLM
+#> [2023-02-10 14:41:26] Initializing ROC-GLM
 #> 
-#> [2023-02-09 14:19:39] Host: Received scores of negative response
-#> [2023-02-09 14:19:39] Receiving negative scores
-#> [2023-02-09 14:19:41] Host: Pushing pooled scores
-#> [2023-02-09 14:19:43] Server: Calculating placement values and parts for ROC-GLM
-#> [2023-02-09 14:19:44] Server: Calculating probit regression to obtain ROC-GLM
-#> [2023-02-09 14:19:46] Deviance of iter1=38.8162
-#> [2023-02-09 14:19:47] Deviance of iter2=48.9408
-#> [2023-02-09 14:19:48] Deviance of iter3=52.5077
-#> [2023-02-09 14:19:50] Deviance of iter4=52.5684
-#> [2023-02-09 14:19:51] Deviance of iter5=52.5684
-#> [2023-02-09 14:19:52] Deviance of iter6=52.5684
-#> [2023-02-09 14:19:52] Host: Finished calculating ROC-GLM
-#> [2023-02-09 14:19:52] Host: Cleaning data on server
-#> [2023-02-09 14:19:53] Host: Calculating AUC and CI
-#> [2023-02-09 14:20:05] Finished!
+#> [2023-02-10 14:41:26] Host: Received scores of negative response
+#> [2023-02-10 14:41:26] Receiving negative scores
+#> [2023-02-10 14:41:27] Host: Pushing pooled scores
+#> [2023-02-10 14:41:29] Server: Calculating placement values and parts for ROC-GLM
+#> [2023-02-10 14:41:30] Server: Calculating probit regression to obtain ROC-GLM
+#> [2023-02-10 14:41:32] Deviance of iter1=38.8162
+#> [2023-02-10 14:41:33] Deviance of iter2=48.9408
+#> [2023-02-10 14:41:34] Deviance of iter3=52.5077
+#> [2023-02-10 14:41:35] Deviance of iter4=52.5684
+#> [2023-02-10 14:41:36] Deviance of iter5=52.5684
+#> [2023-02-10 14:41:37] Deviance of iter6=52.5684
+#> [2023-02-10 14:41:37] Host: Finished calculating ROC-GLM
+#> [2023-02-10 14:41:37] Host: Cleaning data on server
+#> [2023-02-10 14:41:38] Host: Calculating AUC and CI
[…]
