### 2023-02-10 difuture-lmu/dsBinVal#18 @AnthonyOfSeattle
Thank you for the prompt edit to this. I will leave a separate issue with a couple copy edit recommendations for this, but I believe this fulfills my request.
