### 2023-02-12 a8ceab0
Some small modifications to the text and bibliography
