+++ paper/paper.bib
@@ -22,7 +22,7 @@
 @article{prasser2018difuture,
     author = {Prasser, Fabian and Kohlbacher, Oliver and Mansmann, Ulrich and Bauer, Bernhard and Kuhn, Klaus A},
     doi = {10.3414/ME17-02-0022},
-    journal = { Methods Inf Med},
+    journal = { Methods of Information in Medicine},
     title = {Data Integration for Future Medicine (DIFUTURE)},
     volume={57},
     number={S01},
@@ -73,10 +73,11 @@
 }
 
 @Article{pROC,
-  title = {pROC: an open-source package for R and S+ to analyze and compare ROC curves},
+  title = {pROC: an open-source package for {R} and {S+} to analyze and compare ROC curves},
   author = {Xavier Robin and Natacha Turck and Alexandre Hainard and Natalia Tiberti and Frédérique Lisacek and Jean-Charles Sanchez and Markus Müller},
   year = {2011},
   journal = {BMC Bioinformatics},
   volume = {12},
   pages = {77},
+  doi={10.1186/1471-2105-12-77}
 }
