### 2023-02-12 #4545 @csoneson
👋🏻 @schalkdaniel - I went through the submission and made a [PR](https://github.com/difuture-lmu/dsBinVal/pull/19) with some minor suggestions for the text/bibliography. Please check and merge if you agree (you can then ask `@editorialbot` to generate a new proof). Then, the next steps are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
