### 2023-02-13 #4545 @patRyserWelch8
Dear all,

I am delighted you have found a way forward and thank you for your collaboration.  I am more than happy with the statistical part being validated with another opinion. Nonetheless, my close working relationship with DataSHIELD development team has lead to make those comments.  Should the function be added to DataSHIELD, more work may be need.

I would strongly encourage the author to contact DataSHIELD to integrate their work in future release.

Best wishes,

Dr Pat Ryser-Welch

Associate lecturer  - T192, T193, T194, M269, TM351,TM352,TM356, TM358, TM470

TM358 module team

Referred as: Her/She

Let's remember:

Start by doing what's necessary; then do what's possible; and suddenly you are doing the impossible.

 Francis of Assisi

Mankind remains the most extraordinary computer of all....

John F. Kennedy

It does not matter how slowly you go as long as you do not stop.

Confucius

Read me : https://interestingcomputing.com<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Finterestingcomputing.com%2F&data=05%7C01%7Cpat.ryser-welch%40open.ac.uk%7Ca0290154173347370e4a08dadf7fbd62%7C0e2ed45596af4100bed3a8e5fd981685%7C0%7C0%7C638068036962847141%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=%2BL%2BjcezazvrVyggTREYv%2F7%2FpYXlAUPzEdXfxqAzQV28%3D&reserved=0>
