+++ README.md
@@ -138,7 +138,7 @@ load(here::here("Readme_files/mod.rda"))
 
 # Push the model to the DataSHIELD servers:
 pushObject(connections, mod)
-#> [2023-02-10 14:39:27] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
+#> [2023-02-13 12:26:42] Your object is bigger than 1 MB (5.75186157226562 MB). Uploading larger objects may take some time.
 
 # Create a clean data set without NAs:
 ds.completeCases("D", newobj = "D_complete")
@@ -194,27 +194,27 @@ ds.asInteger("D_complete$DIS_DIAB", "truth")
 roc_glm = dsROCGLM(connections, truth_name = "truth", pred_name = "pred",
   dat_name = "D_complete", seed_object = "pred")
 #> 
-#> [2023-02-10 14:41:25] L2 sensitivity is: 0.0015
+#> [2023-02-13 12:29:02] L2 sensitivity is: 0.0015
 #> 
-#> [2023-02-10 14:41:26] Setting: epsilon = 0.2 and delta = 0.1
+#> [2023-02-13 12:29:04] Setting: epsilon = 0.2 and delta = 0.1
 #> 
-#> [2023-02-10 14:41:26] Initializing ROC-GLM
+#> [2023-02-13 12:29:04] Initializing ROC-GLM
 #> 
-#> [2023-02-10 14:41:26] Host: Received scores of negative response
-#> [2023-02-10 14:41:26] Receiving negative scores
-#> [2023-02-10 14:41:27] Host: Pushing pooled scores
-#> [2023-02-10 14:41:29] Server: Calculating placement values and parts for ROC-GLM
-#> [2023-02-10 14:41:30] Server: Calculating probit regression to obtain ROC-GLM
-#> [2023-02-10 14:41:32] Deviance of iter1=38.8162
-#> [2023-02-10 14:41:33] Deviance of iter2=48.9408
-#> [2023-02-10 14:41:34] Deviance of iter3=52.5077
-#> [2023-02-10 14:41:35] Deviance of iter4=52.5684
-#> [2023-02-10 14:41:36] Deviance of iter5=52.5684
-#> [2023-02-10 14:41:37] Deviance of iter6=52.5684
-#> [2023-02-10 14:41:37] Host: Finished calculating ROC-GLM
-#> [2023-02-10 14:41:37] Host: Cleaning data on server
-#> [2023-02-10 14:41:38] Host: Calculating AUC and CI
-#> [2023-02-10 14:41:49] Finished!
+#> [2023-02-13 12:29:04] Host: Received scores of negative response
+#> [2023-02-13 12:29:04] Receiving negative scores
+#> [2023-02-13 12:29:05] Host: Pushing pooled scores
+#> [2023-02-13 12:29:07] Server: Calculating placement values and parts for ROC-GLM
+#> [2023-02-13 12:29:09] Server: Calculating probit regression to obtain ROC-GLM
+#> [2023-02-13 12:29:10] Deviance of iter1=38.8162
+#> [2023-02-13 12:29:12] Deviance of iter2=48.9408
+#> [2023-02-13 12:29:13] Deviance of iter3=52.5077
+#> [2023-02-13 12:29:14] Deviance of iter4=52.5684
+#> [2023-02-13 12:29:15] Deviance of iter5=52.5684
+#> [2023-02-13 12:29:17] Deviance of iter6=52.5684
+#> [2023-02-13 12:29:17] Host: Finished calculating ROC-GLM
+#> [2023-02-13 12:29:17] Host: Cleaning data on server
+#> [2023-02-13 12:29:18] Host: Calculating AUC and CI
+#> [2023-02-13 12:29:30] Finished!
 roc_glm
[…]
