### 2023-02-13 #4545 @schalkdaniel
@csoneson I created a tagged release, the tag is `v.1.0.2`. I also uploaded all files of that release to [Zenodo](https://zenodo.org/record/7634619#.Y-osR9LMJ7g), the DOI is `10.5281/zenodo.7634619`. Please also note that I was asked by one of the authors to add another affiliation, it is listed as:
```
 - name: Munich Center for Machine Learning, Munich, Germany
   index: 4
```
