### 2023-02-21 #4545 @schalkdaniel
@Kevin-Mattheus-Moerman, thanks for your effort. I have changed the license, it should now match the one from the submission.
